Begin to refactor the LSL script implementation for easier editing, reading, and sanity.
