+++ OpenSim/Region/ScriptEngine/Shared/Api/Implementation/LSL_Api.cs
@@ -79,3219 +79,11 @@ namespace OpenSim.Region.ScriptEngine.Shared.Api
     /// <summary>
     /// Contains all LSL ll-functions. This class will be in Default AppDomain.
     /// </summary>
-    public class LSL_Api : MarshalByRefObject, ILSL_Api, IScriptApi
+    public partial class LSL_Api : MarshalByRefObject, ILSL_Api, IScriptApi
     {
-        private static readonly ILog m_log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
+        
 
-        private int m_llRequestAgentDataCacheTimeout;
-        public int LlRequestAgentDataCacheTimeoutMs
-        {
-            get
-            {
-                return 1000 * m_llRequestAgentDataCacheTimeout;
-            }
-            set
-            {
-                m_llRequestAgentDataCacheTimeout = value / 1000;
-            }
-       }
 
-        protected IScriptEngine m_ScriptEngine;
-        protected SceneObjectPart m_host;
-
-        protected UUID RegionScopeID = UUID.Zero;
-        protected string m_regionName = String.Empty;
-        /// <summary>
-        /// The item that hosts this script
-        /// </summary>
-        protected TaskInventoryItem m_item;
-
-        protected bool throwErrorOnNotImplemented = false;
-        protected float m_ScriptDelayFactor = 1.0f;
-        protected float m_Script10mDistance = 10.0f;
-        protected float m_Script10mDistanceSquare = 100.0f;
-        protected float m_MinTimerInterval = 0.5f;
-        protected float m_recoilScaleFactor = 0.0f;
-        protected bool m_AllowGodFunctions;
-
-        protected string m_GetWallclockTimeZone = String.Empty;     // Defaults to UTC
-        protected double m_timer = Util.GetTimeStampMS();
-        protected bool m_waitingForScriptAnswer = false;
-        protected bool m_automaticLinkPermission = false;
-        protected int m_notecardLineReadCharsMax = 255;
-        protected int m_scriptConsoleChannel = 0;
-        protected bool m_scriptConsoleChannelEnabled = false;
-        protected bool m_debuggerSafe = false;
-
-        protected AsyncCommandManager m_AsyncCommands = null;
-        protected IUrlModule m_UrlModule = null;
-        protected IMaterialsModule m_materialsModule = null;
-        protected IEnvironmentModule m_envModule = null;
-        protected IEmailModule m_emailModule = null;
-        protected IUserAccountService m_userAccountService = null;
-        protected IMessageTransferModule m_TransferModule = null;
-
-        protected ExpiringCacheOS<UUID, PresenceInfo> m_PresenceInfoCache = new ExpiringCacheOS<UUID, PresenceInfo>(10000);
-        protected int EMAIL_PAUSE_TIME = 20;  // documented delay value for smtp.
-        protected int m_sleepMsOnSetTexture = 200;
-        protected int m_sleepMsOnSetLinkTexture = 200;
-        protected int m_sleepMsOnScaleTexture = 200;
-        protected int m_sleepMsOnOffsetTexture = 200;
-        protected int m_sleepMsOnRotateTexture = 200;
-        protected int m_sleepMsOnSetPos = 200;
-        protected int m_sleepMsOnSetRot = 200;
-        protected int m_sleepMsOnSetLocalRot = 200;
-        protected int m_sleepMsOnPreloadSound = 1000;
-        protected int m_sleepMsOnMakeExplosion = 100;
-        protected int m_sleepMsOnMakeFountain = 100;
-        protected int m_sleepMsOnMakeSmoke = 100;
-        protected int m_sleepMsOnMakeFire = 100;
-        protected int m_sleepMsOnRezAtRoot = 100;
-        protected int m_sleepMsOnInstantMessage = 2000;
-        protected int m_sleepMsOnEmail = 30000;
-        protected int m_sleepMsOnCreateLink = 1000;
-        protected int m_sleepMsOnGiveInventory = 3000;
-        protected int m_sleepMsOnRequestAgentData = 100;
-        protected int m_sleepMsOnRequestInventoryData = 1000;
-        protected int m_sleepMsOnSetDamage = 5000;
-        protected int m_sleepMsOnTextBox = 1000;
-        protected int m_sleepMsOnAdjustSoundVolume = 100;
-        protected int m_sleepMsOnEjectFromLand = 1000;
-        protected int m_sleepMsOnAddToLandPassList = 100;
-        protected int m_sleepMsOnDialog = 1000;
-        protected int m_sleepMsOnRemoteLoadScript = 3000;
-        protected int m_sleepMsOnRemoteLoadScriptPin = 3000;
-        protected int m_sleepMsOnOpenRemoteDataChannel = 1000;
-        protected int m_sleepMsOnSendRemoteData = 3000;
-        protected int m_sleepMsOnRemoteDataReply = 3000;
-        protected int m_sleepMsOnCloseRemoteDataChannel = 1000;
-        protected int m_sleepMsOnSetPrimitiveParams = 200;
-        protected int m_sleepMsOnSetLinkPrimitiveParams = 200;
-        protected int m_sleepMsOnXorBase64Strings = 300;
-        protected int m_sleepMsOnSetParcelMusicURL = 2000;
-        protected int m_sleepMsOnGetPrimMediaParams = 1000;
-        protected int m_sleepMsOnGetLinkMedia = 1000;
-        protected int m_sleepMsOnSetPrimMediaParams = 1000;
-        protected int m_sleepMsOnSetLinkMedia = 1000;
-        protected int m_sleepMsOnClearPrimMedia = 1000;
-        protected int m_sleepMsOnClearLinkMedia = 1000;
-        protected int m_sleepMsOnRequestSimulatorData = 1000;
-        protected int m_sleepMsOnLoadURL = 1000;
-        protected int m_sleepMsOnParcelMediaCommandList = 2000;
-        protected int m_sleepMsOnParcelMediaQuery = 2000;
-        protected int m_sleepMsOnModPow = 1000;
-        protected int m_sleepMsOnSetPrimURL = 2000;
-        protected int m_sleepMsOnRefreshPrimURL = 20000;
-        protected int m_sleepMsOnMapDestination = 1000;
-        protected int m_sleepMsOnAddToLandBanList = 100;
-        protected int m_sleepMsOnRemoveFromLandPassList = 100;
-        protected int m_sleepMsOnRemoveFromLandBanList = 100;
-        protected int m_sleepMsOnResetLandBanList = 100;
-        protected int m_sleepMsOnResetLandPassList = 100;
-        protected int m_sleepMsOnGetParcelPrimOwners = 2000;
-        protected int m_sleepMsOnGetNumberOfNotecardLines = 100;
-        protected int m_sleepMsOnGetNotecardLine = 100;
-        protected string m_internalObjectHost = "lsl.opensim.local";
-        protected bool m_restrictEmail = false;
-        protected ISoundModule m_SoundModule = null;
-
-        protected float m_avatarHeightCorrection = 0.2f;
-        protected bool m_useSimpleBoxesInGetBoundingBox = false;
-        protected bool m_addStatsInGetBoundingBox = false;
-
-        //LSL Avatar Bounding Box (lABB), lower (1) and upper (2),
-        //standing (Std), Groundsitting (Grs), Sitting (Sit),
-        //along X, Y and Z axes, constants (0) and coefficients (1)
-        protected float m_lABB1StdX0 = -0.275f;
-        protected float m_lABB2StdX0 = 0.275f;
-        protected float m_lABB1StdY0 = -0.35f;
-        protected float m_lABB2StdY0 = 0.35f;
-        protected float m_lABB1StdZ0 = -0.1f;
-        protected float m_lABB1StdZ1 = -0.5f;
-        protected float m_lABB2StdZ0 = 0.1f;
-        protected float m_lABB2StdZ1 = 0.5f;
-        protected float m_lABB1GrsX0 = -0.3875f;
-        protected float m_lABB2GrsX0 = 0.3875f;
-        protected float m_lABB1GrsY0 = -0.5f;
-        protected float m_lABB2GrsY0 = 0.5f;
-        protected float m_lABB1GrsZ0 = -0.05f;
-        protected float m_lABB1GrsZ1 = -0.375f;
-        protected float m_lABB2GrsZ0 = 0.5f;
-        protected float m_lABB2GrsZ1 = 0.0f;
-        protected float m_lABB1SitX0 = -0.5875f;
-        protected float m_lABB2SitX0 = 0.1875f;
-        protected float m_lABB1SitY0 = -0.35f;
-        protected float m_lABB2SitY0 = 0.35f;
-        protected float m_lABB1SitZ0 = -0.35f;
-        protected float m_lABB1SitZ1 = -0.375f;
-        protected float m_lABB2SitZ0 = -0.25f;
-        protected float m_lABB2SitZ1 = 0.25f;
-
-        protected float m_primSafetyCoeffX = 2.414214f;
-        protected float m_primSafetyCoeffY = 2.414214f;
-        protected float m_primSafetyCoeffZ = 1.618034f;
-
-        protected float m_floatToleranceInCastRay = 0.00001f;
-        protected float m_floatTolerance2InCastRay = 0.001f;
-        protected DetailLevel m_primLodInCastRay = DetailLevel.Medium;
-        protected DetailLevel m_sculptLodInCastRay = DetailLevel.Medium;
-        protected DetailLevel m_meshLodInCastRay = DetailLevel.Highest;
-        protected DetailLevel m_avatarLodInCastRay = DetailLevel.Medium;
-        protected int m_maxHitsInCastRay = 16;
-        protected int m_maxHitsPerPrimInCastRay = 16;
-        protected int m_maxHitsPerObjectInCastRay = 16;
-        protected bool m_detectExitsInCastRay = false;
-        protected bool m_doAttachmentsInCastRay = false;
-        protected int m_msThrottleInCastRay = 200;
-        protected int m_msPerRegionInCastRay = 40;
-        protected int m_msPerAvatarInCastRay = 10;
-        protected int m_msMinInCastRay = 2;
-        protected int m_msMaxInCastRay = 40;
-        protected static List<CastRayCall> m_castRayCalls = new List<CastRayCall>();
-        protected bool m_useMeshCacheInCastRay = true;
-        protected static Dictionary<ulong, FacetedMesh> m_cachedMeshes = new Dictionary<ulong, FacetedMesh>();
-
-//        protected Timer m_ShoutSayTimer;
-        protected int m_SayShoutCount = 0;
-        DateTime m_lastSayShoutCheck;
-
-        private int m_whisperdistance = 10;
-        private int m_saydistance = 20;
-        private int m_shoutdistance = 100;
-
-        bool m_disable_underground_movement = true;
-
-        private string m_lsl_shard = "OpenSim";
-        private string m_lsl_user_agent = string.Empty;
-
-        private static readonly Dictionary<string, string> MovementAnimationsForLSL = new Dictionary<string, string>(StringComparer.InvariantCultureIgnoreCase)
-        {
-            {"CROUCH", "Crouching"},
-            {"CROUCHWALK", "CrouchWalking"},
-            {"FALLDOWN", "Falling Down"},
-            {"FLY", "Flying"},
-            {"FLYSLOW", "FlyingSlow"},
-            {"HOVER", "Hovering"},
-            {"HOVER_UP", "Hovering Up"},
-            {"HOVER_DOWN", "Hovering Down"},
-            {"JUMP", "Jumping"},
-            {"LAND", "Landing"},
-            {"PREJUMP", "PreJumping"},
-            {"RUN", "Running"},
-            {"SIT","Sitting"},
-            {"SITGROUND","Sitting on Ground"},
-            {"STAND", "Standing"},
-            {"STANDUP", "Standing Up"},
-            {"STRIDE","Striding"},
-            {"SOFT_LAND", "Soft Landing"},
-            {"TURNLEFT", "Turning Left"},
-            {"TURNRIGHT", "Turning Right"},
-            {"WALK", "Walking"}
-        };
-
-        //llHTTPRequest custom headers use control
-        // true means fatal error,
-        // false means ignore,
-        // missing means allowed
-        private static readonly Dictionary<string,bool> HttpForbiddenHeaders = new Dictionary<string, bool>(StringComparer.InvariantCultureIgnoreCase)
-        {
-            {"Accept", true},
-            {"Accept-Charset", true},
-            {"Accept-CH", false},
-            {"Accept-CH-Lifetime", false},
-            {"Access-Control-Request-Headers", false},
-            {"Access-Control-Request-Method", false},
-            {"Accept-Encoding", false},
-            //{"Accept-Language", false},
-            {"Accept-Patch", false}, // it is server side
-            {"Accept-Post", false}, // it is server side
-            {"Accept-Ranges", false}, // it is server side
-            //{"Age", false},
-            //{"Allow", false},
-            //{"Authorization", false},
-            {"Cache-Control", false},
-            {"Connection", false},
-            {"Content-Length", false},
-            //{"Content-Encoding", false},
-            //{"Content-Location", false},
-            //{"Content-MD5", false},
-            //{"Content-Range", false},
-            {"Content-Type", true},
-            {"Cookie", false},
-            {"Cookie2", false},
-            {"Date", false},
-            {"Device-Memory", false},
-            {"DTN", false},
-            {"Early-Data", false},
-            //{"ETag", false},
-            {"Expect", false},
-            //{"Expires", false},
-            {"Feature-Policy", false},
-            {"From", true},
-            {"Host", true},
-            {"Keep-Alive", false},
-            {"If-Match", false},
-            {"If-Modified-Since", false},
-            {"If-None-Match", false},
-            //{"If-Range", false},
-            {"If-Unmodified-Since", false},
-            //{"Last-Modified", false},
-            //{"Location", false},
-            {"Max-Forwards", false},
-            {"Origin", false},
-            {"Pragma", false},
-            //{"Proxy-Authenticate", false},
-            //{"Proxy-Authorization", false},
-            //{"Range", false},
-            {"Referer", true},
-            //{"Retry-After", false},
-            {"Server", false},
-            {"Set-Cookie", false},
-            {"Set-Cookie2", false},
-            {"TE", true},
-            {"Trailer", true},
-            {"Transfer-Encoding", false},
-            {"Upgrade", true},
-            {"User-Agent", true},
-            {"Vary", false},
-            {"Via", true},
-            {"Viewport-Width", false},
-            {"Warning", false},
-            {"Width", false},
-            //{"WWW-Authenticate", false},
-
-            {"X-Forwarded-For", false},
-            {"X-Forwarded-Host", false},
-            {"X-Forwarded-Proto", false},
-
-            {"x-secondlife-shard", false},
-            {"x-secondlife-object-name", false},
-            {"x-secondlife-object-key", false},
-            {"x-secondlife-region", false},
-            {"x-secondlife-local-position", false},
-            {"x-secondlife-local-velocity", false},
-            {"x-secondlife-local-rotation", false},
-            {"x-secondlife-owner-name", false},
-            {"x-secondlife-owner-key", false},
-        };
-
-        private static readonly HashSet<string> HttpForbiddenInHeaders = new HashSet<string>(StringComparer.InvariantCultureIgnoreCase)
-        {
-            "x-secondlife-shard", "x-secondlife-object-name",  "x-secondlife-object-key",
-            "x-secondlife-region", "x-secondlife-local-position", "x-secondlife-local-velocity",
-            "x-secondlife-local-rotation",  "x-secondlife-owner-name", "x-secondlife-owner-key",
-            "connection", "content-length", "from", "host", "proxy-authorization",
-            "referer", "trailer", "transfer-encoding", "via", "authorization"
-        };
-
-        public void Initialize(IScriptEngine scriptEngine, SceneObjectPart host, TaskInventoryItem item)
-        {
-            m_lastSayShoutCheck = DateTime.UtcNow;
-
-            m_ScriptEngine = scriptEngine;
-            m_host = host;
-            m_item = item;
-            m_debuggerSafe = m_ScriptEngine.Config.GetBoolean("DebuggerSafe", false);
-
-            LoadConfig();
-
-            m_TransferModule = m_ScriptEngine.World.RequestModuleInterface<IMessageTransferModule>();
-            m_UrlModule = m_ScriptEngine.World.RequestModuleInterface<IUrlModule>();
-            m_SoundModule = m_ScriptEngine.World.RequestModuleInterface<ISoundModule>();
-            m_materialsModule = m_ScriptEngine.World.RequestModuleInterface<IMaterialsModule>();
-
-            m_emailModule = m_ScriptEngine.World.RequestModuleInterface<IEmailModule>();
-            m_envModule = m_ScriptEngine.World.RequestModuleInterface< IEnvironmentModule>();
-
-            m_AsyncCommands = new AsyncCommandManager(m_ScriptEngine);
-            m_userAccountService = World.UserAccountService;
-            if(World.RegionInfo != null)
-            {
-                RegionScopeID = World.RegionInfo.ScopeID;
-                m_regionName = World.RegionInfo.RegionName;
-            }
-        }
-
-        /// <summary>
-        /// Load configuration items that affect script, object and run-time behavior. */
-        /// </summary>
-        private void LoadConfig()
-        {
-            LlRequestAgentDataCacheTimeoutMs = 20000;
-
-            IConfig seConfig = m_ScriptEngine.Config;
-
-            if (seConfig != null)
-            {
-                float scriptDistanceFactor = seConfig.GetFloat("ScriptDistanceLimitFactor", 1.0f);
-                m_Script10mDistance = 10.0f * scriptDistanceFactor;
-                m_Script10mDistanceSquare = m_Script10mDistance * m_Script10mDistance;
-
-                m_ScriptDelayFactor = seConfig.GetFloat("ScriptDelayFactor", m_ScriptDelayFactor);
-                m_MinTimerInterval         = seConfig.GetFloat("MinTimerInterval", m_MinTimerInterval);
-                m_automaticLinkPermission  = seConfig.GetBoolean("AutomaticLinkPermission", m_automaticLinkPermission);
-                m_notecardLineReadCharsMax = seConfig.GetInt("NotecardLineReadCharsMax", m_notecardLineReadCharsMax);
-
-                m_GetWallclockTimeZone = seConfig.GetString("GetWallclockTimeZone", m_GetWallclockTimeZone);
-
-                // Rezzing an object with a velocity can create recoil. This feature seems to have been
-                //    removed from recent versions of SL. The code computes recoil (vel*mass) and scales
-                //    it by this factor. May be zero to turn off recoil all together.
-                m_recoilScaleFactor = seConfig.GetFloat("RecoilScaleFactor", m_recoilScaleFactor);
-                m_AllowGodFunctions = seConfig.GetBoolean("AllowGodFunctions", false);
-
-                m_disable_underground_movement = seConfig.GetBoolean("DisableUndergroundMovement", true);
-            }
-
-            if (m_notecardLineReadCharsMax > 65535)
-                m_notecardLineReadCharsMax = 65535;
-
-            // load limits for particular subsystems.
-            IConfigSource seConfigSource = m_ScriptEngine.ConfigSource;
-
-            if (seConfigSource != null)
-            {
-                IConfig netConfig = seConfigSource.Configs["Network"];
-                if (netConfig != null)
-                {
-                    m_lsl_shard = netConfig.GetString("shard", m_lsl_shard);
-                    m_lsl_user_agent = netConfig.GetString("user_agent", m_lsl_user_agent);
-                }
-
-                IConfig lslConfig = seConfigSource.Configs["LL-Functions"];
-                if (lslConfig != null)
-                {
-                    m_restrictEmail = lslConfig.GetBoolean("RestrictEmail", m_restrictEmail);
-                    m_avatarHeightCorrection = lslConfig.GetFloat("AvatarHeightCorrection", m_avatarHeightCorrection);
-                    m_useSimpleBoxesInGetBoundingBox = lslConfig.GetBoolean("UseSimpleBoxesInGetBoundingBox", m_useSimpleBoxesInGetBoundingBox);
-                    m_addStatsInGetBoundingBox = lslConfig.GetBoolean("AddStatsInGetBoundingBox", m_addStatsInGetBoundingBox);
-                    m_lABB1StdX0 = lslConfig.GetFloat("LowerAvatarBoundingBoxStandingXconst", m_lABB1StdX0);
-                    m_lABB2StdX0 = lslConfig.GetFloat("UpperAvatarBoundingBoxStandingXconst", m_lABB2StdX0);
-                    m_lABB1StdY0 = lslConfig.GetFloat("LowerAvatarBoundingBoxStandingYconst", m_lABB1StdY0);
-                    m_lABB2StdY0 = lslConfig.GetFloat("UpperAvatarBoundingBoxStandingYconst", m_lABB2StdY0);
-                    m_lABB1StdZ0 = lslConfig.GetFloat("LowerAvatarBoundingBoxStandingZconst", m_lABB1StdZ0);
-                    m_lABB1StdZ1 = lslConfig.GetFloat("LowerAvatarBoundingBoxStandingZcoeff", m_lABB1StdZ1);
-                    m_lABB2StdZ0 = lslConfig.GetFloat("UpperAvatarBoundingBoxStandingZconst", m_lABB2StdZ0);
-                    m_lABB2StdZ1 = lslConfig.GetFloat("UpperAvatarBoundingBoxStandingZcoeff", m_lABB2StdZ1);
-                    m_lABB1GrsX0 = lslConfig.GetFloat("LowerAvatarBoundingBoxGroundsittingXconst", m_lABB1GrsX0);
-                    m_lABB2GrsX0 = lslConfig.GetFloat("UpperAvatarBoundingBoxGroundsittingXconst", m_lABB2GrsX0);
-                    m_lABB1GrsY0 = lslConfig.GetFloat("LowerAvatarBoundingBoxGroundsittingYconst", m_lABB1GrsY0);
-                    m_lABB2GrsY0 = lslConfig.GetFloat("UpperAvatarBoundingBoxGroundsittingYconst", m_lABB2GrsY0);
-                    m_lABB1GrsZ0 = lslConfig.GetFloat("LowerAvatarBoundingBoxGroundsittingZconst", m_lABB1GrsZ0);
-                    m_lABB1GrsZ1 = lslConfig.GetFloat("LowerAvatarBoundingBoxGroundsittingZcoeff", m_lABB1GrsZ1);
-                    m_lABB2GrsZ0 = lslConfig.GetFloat("UpperAvatarBoundingBoxGroundsittingZconst", m_lABB2GrsZ0);
-                    m_lABB2GrsZ1 = lslConfig.GetFloat("UpperAvatarBoundingBoxGroundsittingZcoeff", m_lABB2GrsZ1);
-                    m_lABB1SitX0 = lslConfig.GetFloat("LowerAvatarBoundingBoxSittingXconst", m_lABB1SitX0);
-                    m_lABB2SitX0 = lslConfig.GetFloat("UpperAvatarBoundingBoxSittingXconst", m_lABB2SitX0);
-                    m_lABB1SitY0 = lslConfig.GetFloat("LowerAvatarBoundingBoxSittingYconst", m_lABB1SitY0);
-                    m_lABB2SitY0 = lslConfig.GetFloat("UpperAvatarBoundingBoxSittingYconst", m_lABB2SitY0);
-                    m_lABB1SitZ0 = lslConfig.GetFloat("LowerAvatarBoundingBoxSittingZconst", m_lABB1SitZ0);
-                    m_lABB1SitZ1 = lslConfig.GetFloat("LowerAvatarBoundingBoxSittingZcoeff", m_lABB1SitZ1);
-                    m_lABB2SitZ0 = lslConfig.GetFloat("UpperAvatarBoundingBoxSittingZconst", m_lABB2SitZ0);
-                    m_lABB2SitZ1 = lslConfig.GetFloat("UpperAvatarBoundingBoxSittingZcoeff", m_lABB2SitZ1);
-                    m_primSafetyCoeffX = lslConfig.GetFloat("PrimBoundingBoxSafetyCoefficientX", m_primSafetyCoeffX);
-                    m_primSafetyCoeffY = lslConfig.GetFloat("PrimBoundingBoxSafetyCoefficientY", m_primSafetyCoeffY);
-                    m_primSafetyCoeffZ = lslConfig.GetFloat("PrimBoundingBoxSafetyCoefficientZ", m_primSafetyCoeffZ);
-                    m_floatToleranceInCastRay = lslConfig.GetFloat("FloatToleranceInLlCastRay", m_floatToleranceInCastRay);
-                    m_floatTolerance2InCastRay = lslConfig.GetFloat("FloatTolerance2InLlCastRay", m_floatTolerance2InCastRay);
-                    m_primLodInCastRay = (DetailLevel)lslConfig.GetInt("PrimDetailLevelInLlCastRay", (int)m_primLodInCastRay);
-                    m_sculptLodInCastRay = (DetailLevel)lslConfig.GetInt("SculptDetailLevelInLlCastRay", (int)m_sculptLodInCastRay);
-                    m_meshLodInCastRay = (DetailLevel)lslConfig.GetInt("MeshDetailLevelInLlCastRay", (int)m_meshLodInCastRay);
-                    m_avatarLodInCastRay = (DetailLevel)lslConfig.GetInt("AvatarDetailLevelInLlCastRay", (int)m_avatarLodInCastRay);
-                    m_maxHitsInCastRay = lslConfig.GetInt("MaxHitsInLlCastRay", m_maxHitsInCastRay);
-                    m_maxHitsPerPrimInCastRay = lslConfig.GetInt("MaxHitsPerPrimInLlCastRay", m_maxHitsPerPrimInCastRay);
-                    m_maxHitsPerObjectInCastRay = lslConfig.GetInt("MaxHitsPerObjectInLlCastRay", m_maxHitsPerObjectInCastRay);
-                    m_detectExitsInCastRay = lslConfig.GetBoolean("DetectExitHitsInLlCastRay", m_detectExitsInCastRay);
-                    m_doAttachmentsInCastRay = lslConfig.GetBoolean("DoAttachmentsInLlCastRay", m_doAttachmentsInCastRay);
-                    m_msThrottleInCastRay = lslConfig.GetInt("ThrottleTimeInMsInLlCastRay", m_msThrottleInCastRay);
-                    m_msPerRegionInCastRay = lslConfig.GetInt("AvailableTimeInMsPerRegionInLlCastRay", m_msPerRegionInCastRay);
-                    m_msPerAvatarInCastRay = lslConfig.GetInt("AvailableTimeInMsPerAvatarInLlCastRay", m_msPerAvatarInCastRay);
-                    m_msMinInCastRay = lslConfig.GetInt("RequiredAvailableTimeInMsInLlCastRay", m_msMinInCastRay);
-                    m_msMaxInCastRay = lslConfig.GetInt("MaximumAvailableTimeInMsInLlCastRay", m_msMaxInCastRay);
-                    m_useMeshCacheInCastRay = lslConfig.GetBoolean("UseMeshCacheInLlCastRay", m_useMeshCacheInCastRay);
-                }
-
-                IConfig smtpConfig = seConfigSource.Configs["SMTP"];
-                if (smtpConfig != null)
-                {
-                    // there's an smtp config, so load in the snooze time.
-                    EMAIL_PAUSE_TIME = smtpConfig.GetInt("email_pause_time", EMAIL_PAUSE_TIME);
-
-                    m_internalObjectHost = smtpConfig.GetString("internal_object_host", m_internalObjectHost);
-                }
-
-                IConfig chatConfig = seConfigSource.Configs["SMTP"];
-                if(chatConfig != null)
-                {
-                    m_whisperdistance = chatConfig.GetInt("whisper_distance", m_whisperdistance);
-                    m_saydistance = chatConfig.GetInt("say_distance", m_saydistance);
-                    m_shoutdistance = chatConfig.GetInt("shout_distance", m_shoutdistance);
-                }
-            }
-            m_sleepMsOnEmail = EMAIL_PAUSE_TIME * 1000;
-        }
-
-        public override Object InitializeLifetimeService()
-        {
-            ILease lease = (ILease)base.InitializeLifetimeService();
-
-            if (lease.CurrentState == LeaseState.Initial)
-            {
-                lease.InitialLeaseTime = TimeSpan.FromMinutes(0);
-//                lease.RenewOnCallTime = TimeSpan.FromSeconds(10.0);
-//                lease.SponsorshipTimeout = TimeSpan.FromMinutes(1.0);
-            }
-            return lease;
-        }
-
-        protected SceneObjectPart MonitoringObject()
-        {
-            UUID m = m_host.ParentGroup.MonitoringObject;
-            if (m.IsZero())
-                return null;
-
-            SceneObjectPart p = m_ScriptEngine.World.GetSceneObjectPart(m);
-            if (p == null)
-                m_host.ParentGroup.MonitoringObject = UUID.Zero;
-
-            return p;
-        }
-
-        protected virtual void ScriptSleep(int delay)
-        {
-            delay = (int)(delay * m_ScriptDelayFactor);
-            if (delay < 10)
-                return;
-
-            Sleep(delay);
-        }
-
-        protected virtual void Sleep(int delay)
-        {
-            if (m_item == null) // Some unit tests don't set this
-                Thread.Sleep(delay);
-            else
-                m_ScriptEngine.SleepScript(m_item.ItemID, delay);
-        }
-
-        /// <summary>
-        /// Check for co-operative termination.
-        /// </summary>
-        /// <param name='delay'>If called with 0, then just the check is performed with no wait.</param>
-
-        public Scene World
-        {
-            get { return m_ScriptEngine.World; }
-        }
-
-        [DebuggerNonUserCode]
-        public void state(string newState)
-        {
-            m_ScriptEngine.SetState(m_item.ItemID, newState);
-        }
-
-        /// <summary>
-        /// Reset the named script. The script must be present
-        /// in the same prim.
-        /// </summary>
-        [DebuggerNonUserCode]
-        public void llResetScript()
-        {
-
-            // We need to tell the URL module, if we hav one, to release
-            // the allocated URLs
-            if (m_UrlModule != null)
-                m_UrlModule.ScriptRemoved(m_item.ItemID);
-
-            m_ScriptEngine.ApiResetScript(m_item.ItemID);
-        }
-
-        public void llResetOtherScript(string name)
-        {
-            UUID item = GetScriptByName(name);
-
-            if (item.IsZero())
-            {
-                Error("llResetOtherScript", "Can't find script '" + name + "'");
-                return;
-            }
-            if(item.Equals(m_item.ItemID))
-                llResetScript();
-            else
-            {
-                m_ScriptEngine.ResetScript(item);
-            }
-        }
-
-        public LSL_Integer llGetScriptState(string name)
-        {
-            UUID item = GetScriptByName(name);
-
-            if (!item.IsZero())
-            {
-                return m_ScriptEngine.GetScriptState(item) ?1:0;
-            }
-
-            Error("llGetScriptState", "Can't find script '" + name + "'");
-
-            // If we didn't find it, then it's safe to
-            // assume it is not running.
-
-            return 0;
-        }
-
-        public void llSetScriptState(string name, int run)
-        {
-            UUID item = GetScriptByName(name);
-
-
-            // These functions are supposed to be robust,
-            // so get the state one step at a time.
-
-            if (!item.IsZero())
-            {
-                m_ScriptEngine.SetScriptState(item, run == 0 ? false : true, item.Equals(m_item.ItemID));
-            }
-            else
-            {
-                Error("llSetScriptState", "Can't find script '" + name + "'");
-            }
-        }
-
-        public List<ScenePresence> GetLinkAvatars(int linkType)
-        {
-            if (m_host == null)
-                return new List<ScenePresence>();
-
-            return GetLinkAvatars(linkType, m_host.ParentGroup);
-
-        }
-
-        public List<ScenePresence> GetLinkAvatars(int linkType, SceneObjectGroup sog)
-        {
-            List<ScenePresence> ret = new List<ScenePresence>();
-            if (sog == null || sog.IsDeleted)
-                return ret;
-
-            List<ScenePresence> avs = sog.GetSittingAvatars();
-            switch (linkType)
-            {
-                case ScriptBaseClass.LINK_SET:
-                    return avs;
-
-                case ScriptBaseClass.LINK_ROOT:
-                    return ret;
-
-                case ScriptBaseClass.LINK_ALL_OTHERS:
-                    return avs;
-
-                case ScriptBaseClass.LINK_ALL_CHILDREN:
-                    return avs;
-
-                case ScriptBaseClass.LINK_THIS:
-                    return ret;
-
-                default:
-                    if (linkType < 0)
-                        return ret;
-
-                    int partCount = sog.GetPartCount();
-
-                    linkType -= partCount;
-                    if (linkType <= 0)
-                    {
-                        return ret;
-                    }
-                    else
-                    {
-                        if (linkType > avs.Count)
-                        {
-                            return ret;
-                        }
-                        else
-                        {
-                            ret.Add(avs[linkType-1]);
-                            return ret;
-                        }
-                    }
-            }
-        }
-
-        /// <summary>
-        /// Get a given link entity from a linkset (linked objects and any sitting avatars).
-        /// </summary>
-        /// <remarks>
-        /// If there are any ScenePresence's in the linkset (i.e. because they are sat upon one of the prims), then
-        /// these are counted as extra entities that correspond to linknums beyond the number of prims in the linkset.
-        /// The ScenePresences receive linknums in the order in which they sat.
-        /// </remarks>
-        /// <returns>
-        /// The link entity.  null if not found.
-        /// </returns>
-        /// <param name='part'></param>
-        /// <param name='linknum'>
-        /// Can be either a non-negative integer or ScriptBaseClass.LINK_THIS (-4).
-        /// If ScriptBaseClass.LINK_THIS then the entity containing the script is returned.
-        /// If the linkset has one entity and a linknum of zero is given, then the single entity is returned.  If any
-        /// positive integer is given in this case then null is returned.
-        /// If the linkset has more than one entity and a linknum greater than zero but equal to or less than the number
-        /// of entities, then the entity which corresponds to that linknum is returned.
-        /// Otherwise, if a positive linknum is given which is greater than the number of entities in the linkset, then
-        /// null is returned.
-        /// </param>
-        public ISceneEntity GetLinkEntity(SceneObjectPart part, int linknum)
-        {
-            if (linknum < 0)
-            {
-                if (linknum == ScriptBaseClass.LINK_THIS)
-                    return part;
-                else
-                    return null;
-            }
-
-            int actualPrimCount = part.ParentGroup.PrimCount;
-            List<ScenePresence> sittingAvatars = part.ParentGroup.GetSittingAvatars();
-            int adjustedPrimCount = actualPrimCount + sittingAvatars.Count;
-
-            // Special case for a single prim.  In this case the linknum is zero.  However, this will not match a single
-            // prim that has any avatars sat upon it (in which case the root prim is link 1).
-            if (linknum == 0)
-            {
-                if (actualPrimCount == 1 && sittingAvatars.Count == 0)
-                    return part;
-
-                return null;
-            }
-            // Special case to handle a single prim with sitting avatars.  GetLinkPart() would only match zero but
-            // here we must match 1 (ScriptBaseClass.LINK_ROOT).
-            else if (linknum == ScriptBaseClass.LINK_ROOT && actualPrimCount == 1)
-            {
-                if (sittingAvatars.Count > 0)
-                    return part.ParentGroup.RootPart;
-                else
-                    return null;
-            }
-            else if (linknum <= adjustedPrimCount)
-            {
-                if (linknum <= actualPrimCount)
-                {
-                    return part.ParentGroup.GetLinkNumPart(linknum);
-                }
-                else
-                {
-                    return sittingAvatars[linknum - actualPrimCount - 1];
-                }
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public List<SceneObjectPart> GetLinkParts(int linkType)
-        {
-            return GetLinkParts(m_host, linkType);
-        }
-
-        public static List<SceneObjectPart> GetLinkParts(SceneObjectPart part, int linkType)
-        {
-            List<SceneObjectPart> ret = new List<SceneObjectPart>();
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return ret;
-
-            switch (linkType)
-            {
-                case ScriptBaseClass.LINK_SET:
-                    return new List<SceneObjectPart>(part.ParentGroup.Parts);
-
-                case ScriptBaseClass.LINK_ROOT:
-                    ret.Add(part.ParentGroup.RootPart);
-                    return ret;
-
-                case ScriptBaseClass.LINK_ALL_OTHERS:
-                    ret = new List<SceneObjectPart>(part.ParentGroup.Parts);
-
-                    if (ret.Contains(part))
-                        ret.Remove(part);
-
-                    return ret;
-
-                case ScriptBaseClass.LINK_ALL_CHILDREN:
-                    ret = new List<SceneObjectPart>(part.ParentGroup.Parts);
-
-                    if (ret.Contains(part.ParentGroup.RootPart))
-                        ret.Remove(part.ParentGroup.RootPart);
-                    return ret;
-
-                case ScriptBaseClass.LINK_THIS:
-                    ret.Add(part);
-                    return ret;
-
-                default:
-                    if (linkType < 0)
-                        return ret;
-
-                    SceneObjectPart target = part.ParentGroup.GetLinkNumPart(linkType);
-                    if (target == null)
-                        return ret;
-                    ret.Add(target);
-                    return ret;
-            }
-        }
-
-        public List<ISceneEntity> GetLinkEntities(int linkType)
-        {
-            return GetLinkEntities(m_host, linkType);
-        }
-
-        public List<ISceneEntity> GetLinkEntities(SceneObjectPart part, int linkType)
-        {
-            List<ISceneEntity> ret;
-
-            switch (linkType)
-            {
-                case ScriptBaseClass.LINK_SET:
-                    return new List<ISceneEntity>(part.ParentGroup.Parts);
-
-                case ScriptBaseClass.LINK_ROOT:
-                    return new List<ISceneEntity>() { part.ParentGroup.RootPart };
-
-                case ScriptBaseClass.LINK_ALL_OTHERS:
-                    ret = new List<ISceneEntity>(part.ParentGroup.Parts);
-
-                    if (ret.Contains(part))
-                        ret.Remove(part);
-
-                    return ret;
-
-                case ScriptBaseClass.LINK_ALL_CHILDREN:
-                    ret = new List<ISceneEntity>(part.ParentGroup.Parts);
-
-                    if (ret.Contains(part.ParentGroup.RootPart))
-                        ret.Remove(part.ParentGroup.RootPart);
-
-                    List<ScenePresence> avs = part.ParentGroup.GetSittingAvatars();
-                    if(avs!= null && avs.Count > 0)
-                        ret.AddRange(avs);
-
-                    return ret;
-
-                case ScriptBaseClass.LINK_THIS:
-                    return new List<ISceneEntity>() { part };
-
-                default:
-                    if (linkType < 0)
-                        return new List<ISceneEntity>();
-
-                    ISceneEntity target = GetLinkEntity(part, linkType);
-                    if (target == null)
-                        return new List<ISceneEntity>();
-
-                    return new List<ISceneEntity>() { target };
-            }
-        }
-
-        //These are the implementations of the various ll-functions used by the LSL scripts.
-        public LSL_Float llSin(double f)
-        {
-            return Math.Sin(f);
-        }
-
-        public LSL_Float llCos(double f)
-        {
-            return Math.Cos(f);
-        }
-
-        public LSL_Float llTan(double f)
-        {
-            return Math.Tan(f);
-        }
-
-        public LSL_Float llAtan2(LSL_Float x, LSL_Float y)
-        {
-            return Math.Atan2(x, y);
-        }
-
-        public LSL_Float llSqrt(double f)
-        {
-            return Math.Sqrt(f);
-        }
-
-        public LSL_Float llPow(double fbase, double fexponent)
-        {
-            return (double)Math.Pow(fbase, fexponent);
-        }
-
-        public LSL_Integer llAbs(LSL_Integer i)
-        {
-            // changed to replicate LSL behaviour whereby minimum int value is returned untouched.
-            if (i == Int32.MinValue)
-                return i;
-            else
-                return Math.Abs(i);
-        }
-
-        public LSL_Float llFabs(double f)
-        {
-            return (double)Math.Abs(f);
-        }
-
-        public LSL_Float llFrand(double mag)
-        {
-            lock (Util.RandomClass)
-            {
-                return Util.RandomClass.NextDouble() * mag;
-            }
-        }
-
-        public LSL_Integer llFloor(double f)
-        {
-            return (int)Math.Floor(f);
-        }
-
-        public LSL_Integer llCeil(double f)
-        {
-            return (int)Math.Ceiling(f);
-        }
-
-        // Xantor 01/May/2008 fixed midpointrounding (2.5 becomes 3.0 instead of 2.0, default = ToEven)
-        public LSL_Integer llRound(double f)
-        {
-            return (int)Math.Round(f, MidpointRounding.AwayFromZero);
-        }
-
-        //This next group are vector operations involving squaring and square root. ckrinke
-        public LSL_Float llVecMag(LSL_Vector v)
-        {
-            return LSL_Vector.Mag(v);
-        }
-
-        public LSL_Vector llVecNorm(LSL_Vector v)
-        {
-            return LSL_Vector.Norm(v);
-        }
-
-        private double VecDist(LSL_Vector a, LSL_Vector b)
-        {
-            double dx = a.x - b.x;
-            double dy = a.y - b.y;
-            double dz = a.z - b.z;
-            return Math.Sqrt(dx * dx + dy * dy + dz * dz);
-        }
-
-        private double VecDistSquare(LSL_Vector a, LSL_Vector b)
-        {
-            double dx = a.x - b.x;
-            double dy = a.y - b.y;
-            double dz = a.z - b.z;
-            return dx * dx + dy * dy + dz * dz;
-        }
-
-        public LSL_Float llVecDist(LSL_Vector a, LSL_Vector b)
-        {
-            return VecDist(a, b);
-        }
-
-        //Now we start getting into quaternions which means sin/cos, matrices and vectors. ckrinke
-
-        // Utility function for llRot2Euler
-
-        public LSL_Vector llRot2Euler(LSL_Rotation q1)
-        {
-            LSL_Vector eul = new LSL_Vector();
-
-            double sqw = q1.s*q1.s;
-            double sqx = q1.x*q1.x;
-            double sqy = q1.z*q1.z;
-            double sqz = q1.y*q1.y;
-            double unit = sqx + sqy + sqz + sqw; // if normalised is one, otherwise is correction factor
-            double test = q1.x*q1.z + q1.y*q1.s;
-            if (test > 0.4999*unit) { // singularity at north pole
-                eul.z = 2 * Math.Atan2(q1.x,q1.s);
-                eul.y = Math.PI/2;
-                eul.x = 0;
-                return eul;
-            }
-            if (test < -0.4999*unit) { // singularity at south pole
-                eul.z = -2 * Math.Atan2(q1.x,q1.s);
-                eul.y = -Math.PI/2;
-                eul.x = 0;
-                return eul;
-            }
-            eul.z = Math.Atan2(2*q1.z*q1.s-2*q1.x*q1.y , sqx - sqy - sqz + sqw);
-            eul.y = Math.Asin(2*test/unit);
-            eul.x = Math.Atan2(2*q1.x*q1.s-2*q1.z*q1.y , -sqx + sqy - sqz + sqw);
-            return eul;
-        }
-
-        public LSL_Rotation llEuler2Rot(LSL_Vector v)
-        {
-
-            double x,y,z,s;
-            v.x *= 0.5;
-            v.y *= 0.5;
-            v.z *= 0.5;
-            double c1 = Math.Cos(v.x);
-            double c2 = Math.Cos(v.y);
-            double c1c2 = c1 * c2;
-            double s1 = Math.Sin(v.x);
-            double s2 = Math.Sin(v.y);
-            double s1s2 = s1 * s2;
-            double c1s2 = c1 * s2;
-            double s1c2 = s1 * c2;
-            double c3 = Math.Cos(v.z);
-            double s3 = Math.Sin(v.z);
-
-            x = s1c2 * c3 + c1s2 * s3;
-            y = c1s2 * c3 - s1c2 * s3;
-            z = s1s2 * c3 + c1c2 * s3;
-            s = c1c2 * c3 - s1s2 * s3;
-
-            return new LSL_Rotation(x, y, z, s);
-        }
-
-        public LSL_Rotation llAxes2Rot(LSL_Vector fwd, LSL_Vector left, LSL_Vector up)
-        {
-            double s;
-            double tr = fwd.x + left.y + up.z + 1.0;
-
-            if (tr >= 1.0)
-            {
-                s = 0.5 / Math.Sqrt(tr);
-                return new LSL_Rotation(
-                        (left.z - up.y) * s,
-                        (up.x - fwd.z) * s,
-                        (fwd.y - left.x) * s,
-                        0.25 / s);
-            }
-            else
-            {
-                double max = (left.y > up.z) ? left.y : up.z;
-
-                if (max < fwd.x)
-                {
-                    s = Math.Sqrt(fwd.x - (left.y + up.z) + 1.0);
-                    double x = s * 0.5;
-                    s = 0.5 / s;
-                    return new LSL_Rotation(
-                            x,
-                            (fwd.y + left.x) * s,
-                            (up.x + fwd.z) * s,
-                            (left.z - up.y) * s);
-                }
-                else if (max == left.y)
-                {
-                    s = Math.Sqrt(left.y - (up.z + fwd.x) + 1.0);
-                    double y = s * 0.5;
-                    s = 0.5 / s;
-                    return new LSL_Rotation(
-                            (fwd.y + left.x) * s,
-                            y,
-                            (left.z + up.y) * s,
-                            (up.x - fwd.z) * s);
-                }
-                else
-                {
-                    s = Math.Sqrt(up.z - (fwd.x + left.y) + 1.0);
-                    double z = s * 0.5;
-                    s = 0.5 / s;
-                    return new LSL_Rotation(
-                            (up.x + fwd.z) * s,
-                            (left.z + up.y) * s,
-                            z,
-                            (fwd.y - left.x) * s);
-                }
-            }
-        }
-
-        public LSL_Vector llRot2Fwd(LSL_Rotation r)
-        {
-
-            double x, y, z, m;
-
-            m = r.x * r.x + r.y * r.y + r.z * r.z + r.s * r.s;
-            // m is always greater than zero
-            // if m is not equal to 1 then Rotation needs to be normalized
-            if (Math.Abs(1.0 - m) > 0.000001) // allow a little slop here for calculation precision
-            {
-                m = 1.0 / Math.Sqrt(m);
-                r.x *= m;
-                r.y *= m;
-                r.z *= m;
-                r.s *= m;
-            }
-
-            // Fast Algebric Calculations instead of Vectors & Quaternions Product
-            x = r.x * r.x - r.y * r.y - r.z * r.z + r.s * r.s;
-            y = 2 * (r.x * r.y + r.z * r.s);
-            z = 2 * (r.x * r.z - r.y * r.s);
-            return (new LSL_Vector(x, y, z));
-        }
-
-        public LSL_Vector llRot2Left(LSL_Rotation r)
-        {
-
-            double x, y, z, m;
-
-            m = r.x * r.x + r.y * r.y + r.z * r.z + r.s * r.s;
-            // m is always greater than zero
-            // if m is not equal to 1 then Rotation needs to be normalized
-            if (Math.Abs(1.0 - m) > 0.000001) // allow a little slop here for calculation precision
-            {
-                m = 1.0 / Math.Sqrt(m);
-                r.x *= m;
-                r.y *= m;
-                r.z *= m;
-                r.s *= m;
-            }
-
-            // Fast Algebric Calculations instead of Vectors & Quaternions Product
-            x = 2 * (r.x * r.y - r.z * r.s);
-            y = -r.x * r.x + r.y * r.y - r.z * r.z + r.s * r.s;
-            z = 2 * (r.x * r.s + r.y * r.z);
-            return (new LSL_Vector(x, y, z));
-        }
-
-        public LSL_Vector llRot2Up(LSL_Rotation r)
-        {
-            double x, y, z, m;
-
-            m = r.x * r.x + r.y * r.y + r.z * r.z + r.s * r.s;
-            // m is always greater than zero
-            // if m is not equal to 1 then Rotation needs to be normalized
-            if (Math.Abs(1.0 - m) > 0.000001) // allow a little slop here for calculation precision
-            {
-                m = 1.0 / Math.Sqrt(m);
-                r.x *= m;
-                r.y *= m;
-                r.z *= m;
-                r.s *= m;
-            }
-
-            // Fast Algebric Calculations instead of Vectors & Quaternions Product
-            x = 2 * (r.x * r.z + r.y * r.s);
-            y = 2 * (-r.x * r.s + r.y * r.z);
-            z = -r.x * r.x - r.y * r.y + r.z * r.z + r.s * r.s;
-            return (new LSL_Vector(x, y, z));
-        }
-
-        public LSL_Rotation llRotBetween(LSL_Vector a, LSL_Vector b)
-        {
-            //A and B should both be normalized
-
-            // This method mimics the 180 errors found in SL
-            // See www.euclideanspace.com... angleBetween
-
-            // Eliminate zero length
-            LSL_Float vec_a_mag = LSL_Vector.MagSquare(a);
-            LSL_Float vec_b_mag = LSL_Vector.MagSquare(b);
-            if (vec_a_mag < 1e-12 ||
-                vec_b_mag < 1e-12)
-            {
-                return new LSL_Rotation(0.0f, 0.0f, 0.0f, 1.0f);
-            }
-
-            // Normalize
-            a = llVecNorm(a);
-            b = llVecNorm(b);
-
-            // Calculate axis and rotation angle
-            LSL_Vector axis = a % b;
-            LSL_Float cos_theta  = a * b;
-
-            // Check if parallel
-            if (cos_theta > 0.99999)
-            {
-                return new LSL_Rotation(0.0f, 0.0f, 0.0f, 1.0f);
-            }
-
-            // Check if anti-parallel
-            else if (cos_theta < -0.99999)
-            {
-                LSL_Vector orthog_axis = new LSL_Vector(1.0, 0.0, 0.0) - (a.x / (a * a) * a);
-                if (LSL_Vector.MagSquare(orthog_axis)  < 1e-12)
-                    orthog_axis = new LSL_Vector(0.0, 0.0, 1.0);
-                return new LSL_Rotation((float)orthog_axis.x, (float)orthog_axis.y, (float)orthog_axis.z, 0.0);
-            }
-            else // other rotation
-            {
-                LSL_Float theta = (LSL_Float)Math.Acos(cos_theta) * 0.5f;
-                axis = llVecNorm(axis);
-                double x, y, z, s, t;
-                s = Math.Cos(theta);
-                t = Math.Sin(theta);
-                x = axis.x * t;
-                y = axis.y * t;
-                z = axis.z * t;
-                return new LSL_Rotation(x,y,z,s);
-            }
-        }
-
-        public void llWhisper(int channelID, string text)
-        {
-            byte[] binText = Utils.StringToBytesNoTerm(text, 1023);
-            World.SimChat(binText,
-                          ChatTypeEnum.Whisper, channelID, m_host.AbsolutePosition, m_host.Name, m_host.UUID, false);
-
-            IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-            if (wComm != null)
-                wComm.DeliverMessage(ChatTypeEnum.Whisper, channelID, m_host.Name, m_host.UUID, Util.UTF8.GetString(binText), m_host.AbsolutePosition);
-        }
-
-        private void CheckSayShoutTime()
-        {
-            DateTime now = DateTime.UtcNow;
-            if ((now - m_lastSayShoutCheck).Ticks > 10000000) // 1sec
-            {
-                m_lastSayShoutCheck = now;
-                m_SayShoutCount = 0;
-            }
-            else
-                m_SayShoutCount++;
-        }
-
-        public void ThrottleSay(int channelID, int timeMs)
-        {
-            if (channelID == 0)
-                CheckSayShoutTime();
-            if (m_SayShoutCount >= 11)
-                ScriptSleep(timeMs);
-        }
-
-        public void llSay(int channelID, string text)
-        {
-
-            if (channelID == 0)
-//                m_SayShoutCount++;
-                CheckSayShoutTime();
-
-            if (m_SayShoutCount >= 11)
-                ScriptSleep(2000);
-
-            if (m_scriptConsoleChannelEnabled && (channelID == m_scriptConsoleChannel))
-            {
-                Console.WriteLine(text);
-            }
-            else
-            {
-                byte[] binText = Utils.StringToBytesNoTerm(text, 1023);
-                World.SimChat(binText,
-                              ChatTypeEnum.Say, channelID, m_host.AbsolutePosition, m_host.Name, m_host.UUID, false);
-
-                IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-                if (wComm != null)
-                    wComm.DeliverMessage(ChatTypeEnum.Say, channelID, m_host.Name, m_host.UUID, Util.UTF8.GetString(binText), m_host.AbsolutePosition);
-            }
-        }
-
-        public void llShout(int channelID, string text)
-        {
-
-            if (channelID == 0)
-//                m_SayShoutCount++;
-                CheckSayShoutTime();
-
-            if (m_SayShoutCount >= 11)
-                ScriptSleep(2000);
-
-            byte[] binText = Utils.StringToBytesNoTerm(text, 1023);
-
-            World.SimChat(binText,
-                          ChatTypeEnum.Shout, channelID, m_host.AbsolutePosition, m_host.Name, m_host.UUID, true);
-
-            IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-            if (wComm != null)
-                wComm.DeliverMessage(ChatTypeEnum.Shout, channelID, m_host.Name, m_host.UUID, Util.UTF8.GetString(binText), m_host.AbsolutePosition);
-        }
-
-        public void llRegionSay(int channelID, string text)
-        {
-            if (channelID == 0)
-            {
-                Error("llRegionSay", "Cannot use on channel 0");
-                return;
-            }
-
-            byte[] binText = Utils.StringToBytesNoTerm(text, 1023);
-
-            // debug channel is also sent to avatars
-            if (channelID == ScriptBaseClass.DEBUG_CHANNEL)
-            {
-                World.SimChat(binText,
-                    ChatTypeEnum.Shout, channelID, m_host.ParentGroup.RootPart.AbsolutePosition, m_host.Name, m_host.UUID, true);
-            }
-
-            IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-            if (wComm != null)
-                wComm.DeliverMessage(ChatTypeEnum.Region, channelID, m_host.Name, m_host.UUID, Util.UTF8.GetString(binText));
-        }
-
-        public void  llRegionSayTo(string target, int channel, string msg)
-        {
-            if (channel == ScriptBaseClass.DEBUG_CHANNEL)
-                return;
-
-            if(UUID.TryParse(target, out UUID TargetID) && TargetID.IsNotZero())
-            {
-                IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-                if (wComm != null)
-                {
-                    if (msg.Length > 1023)
-                        msg = msg.Substring(0, 1023);
-
-                    wComm.DeliverMessageTo(TargetID, channel, m_host.AbsolutePosition, m_host.Name, m_host.UUID, msg);
-                }
-            }
-        }
-
-        public LSL_Integer llListen(int channelID, string name, string ID, string msg)
-        {
-            IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-            if (wComm == null)
-                return -1;
-
-            UUID.TryParse(ID, out UUID keyID);
-            return wComm.Listen(m_item.ItemID, m_host.UUID, channelID, name, keyID, msg);
-        }
-
-        public void llListenControl(int number, int active)
-        {
-            IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-            if (wComm != null)
-                wComm.ListenControl(m_item.ItemID, number, active);
-        }
-
-        public void llListenRemove(int number)
-        {
-            IWorldComm wComm = m_ScriptEngine.World.RequestModuleInterface<IWorldComm>();
-            if (wComm != null)
-                wComm.ListenRemove(m_item.ItemID, number);
-        }
-
-        public void llSensor(string name, string id, int type, double range, double arc)
-        {
-            UUID.TryParse(id, out UUID keyID);
-            m_AsyncCommands.SensorRepeatPlugin.SenseOnce(m_host.LocalId, m_item.ItemID, name, keyID, type, range, arc, m_host);
-       }
-
-        public void llSensorRepeat(string name, string id, int type, double range, double arc, double rate)
-        {
-            UUID.TryParse(id, out UUID keyID);
-            m_AsyncCommands.SensorRepeatPlugin.SetSenseRepeatEvent(m_host.LocalId, m_item.ItemID, name, keyID, type, range, arc, rate, m_host);
-        }
-
-        public void llSensorRemove()
-        {
-            m_AsyncCommands.SensorRepeatPlugin.UnSetSenseRepeaterEvents(m_host.LocalId, m_item.ItemID);
-        }
-
-        public string resolveName(UUID objecUUID)
-        {
-            // try avatar username surname
-            UserAccount account = m_userAccountService.GetUserAccount(RegionScopeID, objecUUID);
-            if (account != null)
-            {
-                string avatarname = account.Name;
-                return avatarname;
-            }
-            // try an scene object
-            SceneObjectPart SOP = World.GetSceneObjectPart(objecUUID);
-            if (SOP != null)
-            {
-                string objectname = SOP.Name;
-                return objectname;
-            }
-
-            World.Entities.TryGetValue(objecUUID, out EntityBase SensedObject);
-
-            if (SensedObject == null)
-            {
-                IGroupsModule groups = World.RequestModuleInterface<IGroupsModule>();
-                if (groups != null)
-                {
-                    GroupRecord gr = groups.GetGroupRecord(objecUUID);
-                    if (gr != null)
-                        return gr.GroupName;
-                }
-                return String.Empty;
-            }
-
-            return SensedObject.Name;
-        }
-
-        public LSL_String llDetectedName(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return String.Empty;
-            return detectedParams.Name;
-        }
-
-        public LSL_Key llDetectedKey(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return String.Empty;
-            return detectedParams.Key.ToString();
-        }
-
-        public LSL_Key llDetectedOwner(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return String.Empty;
-            return detectedParams.Owner.ToString();
-        }
-
-        public LSL_Integer llDetectedType(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return 0;
-            return new LSL_Integer(detectedParams.Type);
-        }
-
-        public LSL_Vector llDetectedPos(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return new LSL_Vector();
-            return detectedParams.Position;
-        }
-
-        public LSL_Vector llDetectedVel(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return new LSL_Vector();
-            return detectedParams.Velocity;
-        }
-
-        public LSL_Vector llDetectedGrab(int number)
-        {
-            DetectParams parms = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (parms == null)
-                return new LSL_Vector(0, 0, 0);
-
-            return parms.OffsetPos;
-        }
-
-        public LSL_Rotation llDetectedRot(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return new LSL_Rotation();
-            return detectedParams.Rotation;
-        }
-
-        public LSL_Integer llDetectedGroup(int number)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (detectedParams == null)
-                return new LSL_Integer(0);
-            if (m_host.GroupID.Equals(detectedParams.Group))
-                return new LSL_Integer(1);
-            return new LSL_Integer(0);
-        }
-
-        public LSL_Integer llDetectedLinkNumber(int number)
-        {
-            DetectParams parms = m_ScriptEngine.GetDetectParams(m_item.ItemID, number);
-            if (parms == null)
-                return new LSL_Integer(0);
-
-            return new LSL_Integer(parms.LinkNum);
-        }
-
-        /// <summary>
-        /// See http://wiki.secondlife.com/wiki/LlDetectedTouchBinormal for details
-        /// </summary>
-        public LSL_Vector llDetectedTouchBinormal(int index)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, index);
-            if (detectedParams == null)
-                return new LSL_Vector();
-            return detectedParams.TouchBinormal;
-        }
-
-        /// <summary>
-        /// See http://wiki.secondlife.com/wiki/LlDetectedTouchFace for details
-        /// </summary>
-        public LSL_Integer llDetectedTouchFace(int index)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, index);
-            if (detectedParams == null)
-                return new LSL_Integer(-1);
-            return new LSL_Integer(detectedParams.TouchFace);
-        }
-
-        /// <summary>
-        /// See http://wiki.secondlife.com/wiki/LlDetectedTouchNormal for details
-        /// </summary>
-        public LSL_Vector llDetectedTouchNormal(int index)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, index);
-            if (detectedParams == null)
-                return new LSL_Vector();
-            return detectedParams.TouchNormal;
-        }
-
-        /// <summary>
-        /// See http://wiki.secondlife.com/wiki/LlDetectedTouchPos for details
-        /// </summary>
-        public LSL_Vector llDetectedTouchPos(int index)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, index);
-            if (detectedParams == null)
-                return new LSL_Vector();
-            return detectedParams.TouchPos;
-        }
-
-        /// <summary>
-        /// See http://wiki.secondlife.com/wiki/LlDetectedTouchST for details
-        /// </summary>
-        public LSL_Vector llDetectedTouchST(int index)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, index);
-            if (detectedParams == null)
-                return new LSL_Vector(-1.0, -1.0, 0.0);
-            return detectedParams.TouchST;
-        }
-
-        /// <summary>
-        /// See http://wiki.secondlife.com/wiki/LlDetectedTouchUV for details
-        /// </summary>
-        public LSL_Vector llDetectedTouchUV(int index)
-        {
-            DetectParams detectedParams = m_ScriptEngine.GetDetectParams(m_item.ItemID, index);
-            if (detectedParams == null)
-                return new LSL_Vector(-1.0, -1.0, 0.0);
-            return detectedParams.TouchUV;
-        }
-
-        [DebuggerNonUserCode]
-        public virtual void llDie()
-        {
-            if (!m_host.ParentGroup.IsAttachment) throw new SelfDeleteException();
-        }
-
-        public LSL_Float llGround(LSL_Vector offset)
-        {
-            Vector3 pos = m_host.GetWorldPosition() + (Vector3)offset;
-
-            //Get the slope normal.  This gives us the equation of the plane tangent to the slope.
-            LSL_Vector vsn = llGroundNormal(offset);
-
-            // Clamp to valid position
-            if (pos.X < 0)
-                pos.X = 0;
-            else if (pos.X >= World.Heightmap.Width)
-                pos.X = World.Heightmap.Width - 1;
-            if (pos.Y < 0)
-                pos.Y = 0;
-            else if (pos.Y >= World.Heightmap.Height)
-                pos.Y = World.Heightmap.Height - 1;
-
-            //Get the height for the integer coordinates from the Heightmap
-            float baseheight = (float)World.Heightmap[(int)pos.X, (int)pos.Y];
-
-            //Calculate the difference between the actual coordinates and the integer coordinates
-            float xdiff = pos.X - (float)((int)pos.X);
-            float ydiff = pos.Y - (float)((int)pos.Y);
-
-            //Use the equation of the tangent plane to adjust the height to account for slope
-
-            return (((vsn.x * xdiff) + (vsn.y * ydiff)) / (-1 * vsn.z)) + baseheight;
-        }
-
-        public LSL_Float llCloud(LSL_Vector offset)
-        {
-            float cloudCover = 0f;
-            ICloudModule module = World.RequestModuleInterface<ICloudModule>();
-            if (module != null)
-            {
-                Vector3 pos = m_host.GetWorldPosition();
-                int x = (int)(pos.X + offset.x);
-                int y = (int)(pos.Y + offset.y);
-
-                cloudCover = module.CloudCover(x, y, 0);
-
-            }
-            return cloudCover;
-        }
-
-        public LSL_Vector llWind(LSL_Vector offset)
-        {
-            LSL_Vector wind = new LSL_Vector(0, 0, 0);
-            IWindModule module = World.RequestModuleInterface<IWindModule>();
-            if (module != null)
-            {
-                Vector3 pos = m_host.GetWorldPosition();
-                int x = (int)(pos.X + offset.x);
-                int y = (int)(pos.Y + offset.y);
-
-                Vector3 windSpeed = module.WindSpeed(x, y, 0);
-
-                wind.x = windSpeed.X;
-                wind.y = windSpeed.Y;
-            }
-            return wind;
-        }
-
-        public void llSetStatus(int status, int value)
-        {
-            if (m_host == null || m_host.ParentGroup == null || m_host.ParentGroup.IsDeleted)
-                return;
-
-            int statusrotationaxis = 0;
-
-            if ((status & ScriptBaseClass.STATUS_PHYSICS) == ScriptBaseClass.STATUS_PHYSICS)
-            {
-                if (value != 0)
-                {
-                    SceneObjectGroup group = m_host.ParentGroup;
-                    bool allow = true;
-
-                    int maxprims = World.m_linksetPhysCapacity;
-                    bool checkShape = (maxprims > 0 && group.PrimCount > maxprims);
-
-                    foreach (SceneObjectPart part in group.Parts)
-                    {
-                        if (part.PhysicsShapeType == (byte)PhysicsShapeType.None)
-                            continue;
-
-                        if (part.Scale.X > World.m_maxPhys || part.Scale.Y > World.m_maxPhys || part.Scale.Z > World.m_maxPhys)
-                        {
-                            allow = false;
-                            break;
-                        }
-                        if (checkShape)
-                        {
-                            if (--maxprims < 0)
-                            {
-                                allow = false;
-                                break;
-                            }
-                        }
-                    }
-
-                    if (!allow)
-                        return;
-
-                    if (m_host.ParentGroup.RootPart.PhysActor != null &&
-                        m_host.ParentGroup.RootPart.PhysActor.IsPhysical)
-                        return;
-
-                    m_host.ScriptSetPhysicsStatus(true);
-                }
-                else
-                {
-                    m_host.ScriptSetPhysicsStatus(false);
-                }
-            }
-
-            if ((status & ScriptBaseClass.STATUS_PHANTOM) == ScriptBaseClass.STATUS_PHANTOM)
-            {
-                m_host.ParentGroup.ScriptSetPhantomStatus(value != 0);
-            }
-
-            if ((status & ScriptBaseClass.STATUS_CAST_SHADOWS) == ScriptBaseClass.STATUS_CAST_SHADOWS)
-            {
-                m_host.AddFlag(PrimFlags.CastShadows);
-            }
-
-            if ((status & ScriptBaseClass.STATUS_ROTATE_X) == ScriptBaseClass.STATUS_ROTATE_X)
-            {
-                statusrotationaxis |= ScriptBaseClass.STATUS_ROTATE_X;
-            }
-
-            if ((status & ScriptBaseClass.STATUS_ROTATE_Y) == ScriptBaseClass.STATUS_ROTATE_Y)
-            {
-                statusrotationaxis |= ScriptBaseClass.STATUS_ROTATE_Y;
-            }
-
-            if ((status & ScriptBaseClass.STATUS_ROTATE_Z) == ScriptBaseClass.STATUS_ROTATE_Z)
-            {
-                statusrotationaxis |= ScriptBaseClass.STATUS_ROTATE_Z;
-            }
-
-            if ((status & ScriptBaseClass.STATUS_BLOCK_GRAB) == ScriptBaseClass.STATUS_BLOCK_GRAB)
-                m_host.BlockGrab = value != 0;
-
-            if ((status & ScriptBaseClass.STATUS_BLOCK_GRAB_OBJECT) == ScriptBaseClass.STATUS_BLOCK_GRAB_OBJECT)
-                m_host.ParentGroup.BlockGrabOverride = value != 0;
-
-            if ((status & ScriptBaseClass.STATUS_DIE_AT_EDGE) == ScriptBaseClass.STATUS_DIE_AT_EDGE)
-            {
-                if (value != 0)
-                    m_host.SetDieAtEdge(true);
-                else
-                    m_host.SetDieAtEdge(false);
-            }
-
-            if ((status & ScriptBaseClass.STATUS_RETURN_AT_EDGE) == ScriptBaseClass.STATUS_RETURN_AT_EDGE)
-            {
-                if (value != 0)
-                    m_host.SetReturnAtEdge(true);
-                else
-                    m_host.SetReturnAtEdge(false);
-            }
-
-            if ((status & ScriptBaseClass.STATUS_SANDBOX) == ScriptBaseClass.STATUS_SANDBOX)
-            {
-                if (value != 0)
-                    m_host.SetStatusSandbox(true);
-                else
-                    m_host.SetStatusSandbox(false);
-            }
-
-            if (statusrotationaxis != 0)
-            {
-                m_host.SetAxisRotation(statusrotationaxis, value);
-            }
-        }
-
-        private bool IsPhysical()
-        {
-            return ((m_host.GetEffectiveObjectFlags() & (uint)PrimFlags.Physics) == (uint)PrimFlags.Physics);
-        }
-
-        public LSL_Integer llGetStatus(int status)
-        {
-            // m_log.Debug(m_host.ToString() + " status is " + m_host.GetEffectiveObjectFlags().ToString());
-            switch (status)
-            {
-                case ScriptBaseClass.STATUS_PHYSICS:
-                    return IsPhysical() ? 1 : 0;
-
-                case ScriptBaseClass.STATUS_PHANTOM:
-                    if ((m_host.GetEffectiveObjectFlags() & (uint)PrimFlags.Phantom) == (uint)PrimFlags.Phantom)
-                    {
-                        return 1;
-                    }
-                    return 0;
-
-                case ScriptBaseClass.STATUS_CAST_SHADOWS:
-                    if ((m_host.GetEffectiveObjectFlags() & (uint)PrimFlags.CastShadows) == (uint)PrimFlags.CastShadows)
-                    {
-                        return 1;
-                    }
-                    return 0;
-
-                case ScriptBaseClass.STATUS_BLOCK_GRAB:
-                    return m_host.BlockGrab ? 1 : 0;
-
-                case ScriptBaseClass.STATUS_BLOCK_GRAB_OBJECT:
-                    return m_host.ParentGroup.BlockGrabOverride ? 1 : 0;
-
-                case ScriptBaseClass.STATUS_DIE_AT_EDGE:
-                    if (m_host.GetDieAtEdge())
-                        return 1;
-                    else
-                        return 0;
-
-                case ScriptBaseClass.STATUS_RETURN_AT_EDGE:
-                    if (m_host.GetReturnAtEdge())
-                        return 1;
-                    else
-                        return 0;
-
-                case ScriptBaseClass.STATUS_ROTATE_X:
-                    // if (m_host.GetAxisRotation(2) != 0)
-                    if (m_host.GetAxisRotation((int)SceneObjectGroup.axisSelect.STATUS_ROTATE_X) != 0)
-                        return 1;
-                    else
-                        return 0;
-
-                case ScriptBaseClass.STATUS_ROTATE_Y:
-                    if (m_host.GetAxisRotation((int)SceneObjectGroup.axisSelect.STATUS_ROTATE_Y) != 0)
-                        return 1;
-                    else
-                        return 0;
-
-                case ScriptBaseClass.STATUS_ROTATE_Z:
-                    if (m_host.GetAxisRotation((int)SceneObjectGroup.axisSelect.STATUS_ROTATE_Z) != 0)
-                        return 1;
-                    else
-                        return 0;
-
-                case ScriptBaseClass.STATUS_SANDBOX:
-                    if (m_host.GetStatusSandbox())
-                        return 1;
-                    else
-                        return 0;
-            }
-            return 0;
-        }
-
-        public LSL_Integer llScaleByFactor(double scaling_factor)
-        {
-            SceneObjectGroup group = m_host.ParentGroup;
-
-            if(scaling_factor < 1e-6)
-                return ScriptBaseClass.FALSE;
-            if(scaling_factor > 1e6)
-                return ScriptBaseClass.FALSE;
-
-            if (group == null || group.IsDeleted || group.inTransit)
-                return ScriptBaseClass.FALSE;
-
-            if (group.RootPart.PhysActor != null && group.RootPart.PhysActor.IsPhysical)
-                return ScriptBaseClass.FALSE;
-
-            if (group.RootPart.KeyframeMotion != null)
-                return ScriptBaseClass.FALSE;
-
-            if(group.GroupResize(scaling_factor))
-                return ScriptBaseClass.TRUE;
-            else
-                return ScriptBaseClass.FALSE;
-        }
-
-        public LSL_Float llGetMaxScaleFactor()
-        {
-            SceneObjectGroup group = m_host.ParentGroup;
-
-            if (group == null || group.IsDeleted || group.inTransit)
-                return 1.0f;
-
-            return (LSL_Float)group.GetMaxGroupResizeScale();
-        }
-
-        public LSL_Float llGetMinScaleFactor()
-        {
-            SceneObjectGroup group = m_host.ParentGroup;
-
-            if (group == null || group.IsDeleted || group.inTransit)
-                return 1.0f;
-
-            return (LSL_Float)group.GetMinGroupResizeScale();
-        }
-
-        public void llSetScale(LSL_Vector scale)
-        {
-            SetScale(m_host, scale);
-        }
-
-        protected void SetScale(SceneObjectPart part, LSL_Vector scale)
-        {
-            // TODO: this needs to trigger a persistance save as well
-            if (part == null || part.ParentGroup.IsDeleted)
-                return;
-
-            // First we need to check whether or not we need to clamp the size of a physics-enabled prim
-            PhysicsActor pa = part.ParentGroup.RootPart.PhysActor;
-            if (pa != null && pa.IsPhysical)
-            {
-                scale.x = Math.Max(World.m_minPhys, Math.Min(World.m_maxPhys, scale.x));
-                scale.y = Math.Max(World.m_minPhys, Math.Min(World.m_maxPhys, scale.y));
-                scale.z = Math.Max(World.m_minPhys, Math.Min(World.m_maxPhys, scale.z));
-            }
-            else
-            {
-                // If not physical, then we clamp the scale to the non-physical min/max
-                scale.x = Math.Max(World.m_minNonphys, Math.Min(World.m_maxNonphys, scale.x));
-                scale.y = Math.Max(World.m_minNonphys, Math.Min(World.m_maxNonphys, scale.y));
-                scale.z = Math.Max(World.m_minNonphys, Math.Min(World.m_maxNonphys, scale.z));
-            }
-
-            Vector3 tmp = part.Scale;
-            tmp.X = (float)scale.x;
-            tmp.Y = (float)scale.y;
-            tmp.Z = (float)scale.z;
-            part.Scale = tmp;
-            part.ParentGroup.HasGroupChanged = true;
-            part.SendFullUpdateToAllClients();
-        }
-
-        public LSL_Vector llGetScale()
-        {
-            return new LSL_Vector(m_host.Scale.X, m_host.Scale.Y, m_host.Scale.Z);
-        }
-
-        public void llSetClickAction(int action)
-        {
-            m_host.ClickAction = (byte)action;
-            m_host.ParentGroup.HasGroupChanged = true;
-            m_host.ScheduleFullUpdate();
-            return;
-        }
-
-        public void llSetColor(LSL_Vector color, int face)
-        {
-
-            SetColor(m_host, color, face);
-        }
-
-        protected void SetColor(SceneObjectPart part, LSL_Vector color, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            m_host.SetFaceColorAlpha(face, color, null);
-        }
-
-        public void llSetContentType(LSL_Key reqid, LSL_Integer type)
-        {
-
-            if (m_UrlModule == null)
-                return;
-
-            if(!UUID.TryParse(reqid, out UUID id) || id.IsZero())
-                return;
-
-            // Make sure the content type is text/plain to start with
-            m_UrlModule.HttpContentType(id, "text/plain");
-
-            // Is the object owner online and in the region
-            ScenePresence agent = World.GetScenePresence(m_host.ParentGroup.OwnerID);
-            if (agent == null || agent.IsChildAgent || agent.IsDeleted)
-                return;  // Fail if the owner is not in the same region
-
-            // Is it the embeded browser?
-            string userAgent = m_UrlModule.GetHttpHeader(id, "user-agent");
-            if(string.IsNullOrEmpty(userAgent))
-                return;
-
-            if (userAgent.IndexOf("SecondLife") < 0)
-                return; // Not the embedded browser
-
-            // Use the IP address of the client and check against the request
-            // seperate logins from the same IP will allow all of them to get non-text/plain as long
-            // as the owner is in the region. Same as SL!
-            string logonFromIPAddress = agent.ControllingClient.RemoteEndPoint.Address.ToString();
-            if (string.IsNullOrEmpty(logonFromIPAddress))
-                return;
-
-            string requestFromIPAddress = m_UrlModule.GetHttpHeader(id, "x-remote-ip");
-            //m_log.Debug("IP from header='" + requestFromIPAddress + "' IP from endpoint='" + logonFromIPAddress + "'");
-            if (requestFromIPAddress == null)
-                return;
-
-            requestFromIPAddress = requestFromIPAddress.Trim();
-
-            // If the request isnt from the same IP address then the request cannot be from the owner
-            if (!requestFromIPAddress.Equals(logonFromIPAddress))
-                return;
-
-            switch (type)
-            {
-                case ScriptBaseClass.CONTENT_TYPE_HTML:
-                    m_UrlModule.HttpContentType(id, "text/html");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_XML:
-                    m_UrlModule.HttpContentType(id, "application/xml");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_XHTML:
-                    m_UrlModule.HttpContentType(id, "application/xhtml+xml");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_ATOM:
-                    m_UrlModule.HttpContentType(id, "application/atom+xml");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_JSON:
-                    m_UrlModule.HttpContentType(id, "application/json");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_LLSD:
-                    m_UrlModule.HttpContentType(id, "application/llsd+xml");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_FORM:
-                    m_UrlModule.HttpContentType(id, "application/x-www-form-urlencoded");
-                    break;
-                case ScriptBaseClass.CONTENT_TYPE_RSS:
-                    m_UrlModule.HttpContentType(id, "application/rss+xml");
-                    break;
-                default:
-                    m_UrlModule.HttpContentType(id, "text/plain");
-                    break;
-            }
-        }
-
-        public void SetTexGen(SceneObjectPart part, int face,int style)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            MappingType textype;
-            textype = MappingType.Default;
-            if (style == ScriptBaseClass.PRIM_TEXGEN_PLANAR)
-                textype = MappingType.Planar;
-
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                tex.CreateFace((uint) face);
-                tex.FaceTextures[face].TexMapType = textype;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            else if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (uint i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].TexMapType = textype;
-                    }
-                }
-                tex.DefaultTexture.TexMapType = textype;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public void SetGlow(SceneObjectPart part, int face, float glow)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                tex.CreateFace((uint) face);
-                tex.FaceTextures[face].Glow = glow;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            else if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (uint i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].Glow = glow;
-                    }
-                }
-                tex.DefaultTexture.Glow = glow;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public void SetShiny(SceneObjectPart part, int face, int shiny, Bumpiness bump)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Shininess sval = new Shininess();
-
-            switch (shiny)
-            {
-            case 0:
-                sval = Shininess.None;
-                break;
-            case 1:
-                sval = Shininess.Low;
-                break;
-            case 2:
-                sval = Shininess.Medium;
-                break;
-            case 3:
-                sval = Shininess.High;
-                break;
-            default:
-                sval = Shininess.None;
-                break;
-            }
-
-            int nsides = GetNumberOfSides(part);
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            if (face >= 0 && face < nsides)
-            {
-                tex.CreateFace((uint) face);
-                tex.FaceTextures[face].Shiny = sval;
-                tex.FaceTextures[face].Bump = bump;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            else if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (uint i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].Shiny = sval;
-                        tex.FaceTextures[i].Bump = bump;
-                    }
-                }
-                tex.DefaultTexture.Shiny = sval;
-                tex.DefaultTexture.Bump = bump;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public void SetFullBright(SceneObjectPart part, int face, bool bright)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-             int nsides = GetNumberOfSides(part);
-             Primitive.TextureEntry tex = part.Shape.Textures;
-             if (face >= 0 && face < nsides)
-             {
-                 tex.CreateFace((uint) face);
-                 tex.FaceTextures[face].Fullbright = bright;
-                 part.UpdateTextureEntry(tex);
-                 return;
-             }
-             else if (face == ScriptBaseClass.ALL_SIDES)
-             {
-                tex.DefaultTexture.Fullbright = bright;
-                for (uint i = 0; i < nsides; i++)
-                 {
-                    if(tex.FaceTextures[i] != null)
-                        tex.FaceTextures[i].Fullbright = bright;
-                 }
-                 part.UpdateTextureEntry(tex);
-                 return;
-             }
-         }
-
-        public LSL_Float llGetAlpha(int face)
-        {
-
-            return GetAlpha(m_host, face);
-        }
-
-        protected LSL_Float GetAlpha(SceneObjectPart part, int face)
-        {
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-            if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                int i;
-                double sum = 0.0;
-                for (i = 0 ; i < nsides; i++)
-                    sum += (double)tex.GetFace((uint)i).RGBA.A;
-                return sum;
-            }
-            if (face >= 0 && face < nsides)
-            {
-                return (double)tex.GetFace((uint)face).RGBA.A;
-            }
-            return 0.0;
-        }
-
-        public void llSetAlpha(double alpha, int face)
-        {
-
-            SetAlpha(m_host, alpha, face);
-        }
-
-        public void llSetLinkAlpha(int linknumber, double alpha, int face)
-        {
-
-            List<SceneObjectPart> parts = GetLinkParts(linknumber);
-            if (parts.Count > 0)
-            {
-                try
-                {
-                    foreach (SceneObjectPart part in parts)
-                        SetAlpha(part, alpha, face);
-                }
-                finally { }
-            }
-        }
-
-        protected void SetAlpha(SceneObjectPart part, double alpha, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-            Color4 texcolor;
-
-            if (face >= 0 && face < nsides)
-            {
-                texcolor = tex.CreateFace((uint)face).RGBA;
-                texcolor.A = Utils.Clamp((float)alpha, 0.0f, 1.0f);
-                tex.FaceTextures[face].RGBA = texcolor;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            else if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (int i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        texcolor = tex.FaceTextures[i].RGBA;
-                        texcolor.A = Utils.Clamp((float)alpha, 0.0f, 1.0f);
-                        tex.FaceTextures[i].RGBA = texcolor;
-                    }
-                }
-
-                // In some cases, the default texture can be null, eg when every face
-                // has a unique texture
-                if (tex.DefaultTexture != null)
-                {
-                    texcolor = tex.DefaultTexture.RGBA;
-                    texcolor.A = Utils.Clamp((float)alpha, 0.0f, 1.0f);
-                    tex.DefaultTexture.RGBA = texcolor;
-                }
-
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        /// <summary>
-        /// Set flexi parameters of a part.
-        ///
-        /// FIXME: Much of this code should probably be within the part itself.
-        /// </summary>
-        /// <param name="part"></param>
-        /// <param name="flexi"></param>
-        /// <param name="softness"></param>
-        /// <param name="gravity"></param>
-        /// <param name="friction"></param>
-        /// <param name="wind"></param>
-        /// <param name="tension"></param>
-        /// <param name="Force"></param>
-        protected void SetFlexi(SceneObjectPart part, bool flexi, int softness, float gravity, float friction,
-            float wind, float tension, LSL_Vector Force)
-        {
-            if (part == null)
-                return;
-            SceneObjectGroup sog = part.ParentGroup;
-
-            if(sog == null || sog.IsDeleted || sog.inTransit)
-                return;
-
-            PrimitiveBaseShape pbs = part.Shape;
-            pbs.FlexiSoftness = softness;
-            pbs.FlexiGravity = gravity;
-            pbs.FlexiDrag = friction;
-            pbs.FlexiWind = wind;
-            pbs.FlexiTension = tension;
-            pbs.FlexiForceX = (float)Force.x;
-            pbs.FlexiForceY = (float)Force.y;
-            pbs.FlexiForceZ = (float)Force.z;
-
-            pbs.FlexiEntry = flexi;
-
-            if (!pbs.SculptEntry && (pbs.PathCurve == (byte)Extrusion.Straight || pbs.PathCurve == (byte)Extrusion.Flexible))
-            {
-                if(flexi)
-                {
-                    pbs.PathCurve = (byte)Extrusion.Flexible;
-                    if(!sog.IsPhantom)
-                    {
-                        sog.ScriptSetPhantomStatus(true);
-                        return;
-                    }
-                }
-                else
-                {
-                    // Other values not set, they do not seem to be sent to the viewer
-                    // Setting PathCurve appears to be what actually toggles the check box and turns Flexi on and off
-                    pbs.PathCurve = (byte)Extrusion.Straight;
-                }
-            }
-            part.ParentGroup.HasGroupChanged = true;
-            part.ScheduleFullUpdate();
-        }
-
-        /// <summary>
-        /// Set a light point on a part
-        /// </summary>
-        /// FIXME: Much of this code should probably be in SceneObjectGroup
-        ///
-        /// <param name="part"></param>
-        /// <param name="light"></param>
-        /// <param name="color"></param>
-        /// <param name="intensity"></param>
-        /// <param name="radius"></param>
-        /// <param name="falloff"></param>
-        protected void SetPointLight(SceneObjectPart part, bool light, LSL_Vector color, float intensity, float radius, float falloff)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            PrimitiveBaseShape pbs = part.Shape;
-
-            if (light)
-            {
-                pbs.LightEntry = true;
-                pbs.LightColorR = Utils.Clamp((float)color.x, 0.0f, 1.0f);
-                pbs.LightColorG = Utils.Clamp((float)color.y, 0.0f, 1.0f);
-                pbs.LightColorB = Utils.Clamp((float)color.z, 0.0f, 1.0f);
-                pbs.LightIntensity = Utils.Clamp(intensity, 0.0f, 1.0f);
-                pbs.LightRadius = Utils.Clamp(radius, 0.1f, 20.0f);
-                pbs.LightFalloff = Utils.Clamp(falloff, 0.01f, 2.0f);
-            }
-            else
-            {
-                pbs.LightEntry = false;
-            }
-
-            part.ParentGroup.HasGroupChanged = true;
-            part.ScheduleFullUpdate();
-        }
-
-        public LSL_Vector llGetColor(int face)
-        {
-            return GetColor(m_host, face);
-        }
-
-        protected LSL_Vector GetColor(SceneObjectPart part, int face)
-        {
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            Color4 texcolor;
-            LSL_Vector rgb = new LSL_Vector();
-            int nsides = GetNumberOfSides(part);
-
-            if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                int i;
-                for (i = 0; i < nsides; i++)
-                {
-                    texcolor = tex.GetFace((uint)i).RGBA;
-                    rgb.x += texcolor.R;
-                    rgb.y += texcolor.G;
-                    rgb.z += texcolor.B;
-                }
-
-                float invnsides = 1.0f / (float)nsides;
-
-                rgb.x *= invnsides;
-                rgb.y *= invnsides;
-                rgb.z *= invnsides;
-
-                return rgb;
-            }
-            if (face >= 0 && face < nsides)
-            {
-                texcolor = tex.GetFace((uint)face).RGBA;
-                rgb.x = texcolor.R;
-                rgb.y = texcolor.G;
-                rgb.z = texcolor.B;
-
-                return rgb;
-            }
-            else
-            {
-                return new LSL_Vector();
-            }
-        }
-
-        public void llSetTexture(string texture, int face)
-        {
-            SetTexture(m_host, texture, face);
-            ScriptSleep(m_sleepMsOnSetTexture);
-        }
-
-        public void llSetLinkTexture(int linknumber, string texture, int face)
-        {
-
-            List<SceneObjectPart> parts = GetLinkParts(linknumber);
-            if (parts.Count > 0)
-            {
-                try
-                {
-                    foreach (SceneObjectPart part in parts)
-                        SetTexture(part, texture, face);
-                }
-                finally { }
-             }
-            ScriptSleep(m_sleepMsOnSetLinkTexture);
-        }
-
-        protected void SetTextureParams(SceneObjectPart part, string texture, double scaleU, double ScaleV,
-                    double offsetU, double offsetV, double rotation, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            UUID textureID = new UUID();
-            bool dotexture = true;
-            if(String.IsNullOrEmpty(texture) || texture == ScriptBaseClass.NULL_KEY)
-                dotexture = false;
-            else
-            {
-                textureID = ScriptUtils.GetAssetIdFromItemName(m_host, texture, (int)AssetType.Texture);
-                if (textureID.IsZero())
-                {
-                    if (!UUID.TryParse(texture, out textureID))
-                        dotexture = false;
-                }
-            }
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                Primitive.TextureEntryFace texface = tex.CreateFace((uint)face);
-                if (dotexture)
-                    texface.TextureID = textureID;
-                texface.RepeatU = (float)scaleU;
-                texface.RepeatV = (float)ScaleV;
-                texface.OffsetU = (float)offsetU;
-                texface.OffsetV = (float)offsetV;
-                texface.Rotation = (float)rotation;
-                tex.FaceTextures[face] = texface;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            else if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (uint i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        if (dotexture)
-                            tex.FaceTextures[i].TextureID = textureID;
-                        tex.FaceTextures[i].RepeatU = (float)scaleU;
-                        tex.FaceTextures[i].RepeatV = (float)ScaleV;
-                        tex.FaceTextures[i].OffsetU = (float)offsetU;
-                        tex.FaceTextures[i].OffsetV = (float)offsetV;
-                        tex.FaceTextures[i].Rotation = (float)rotation;
-                    }
-                }
-                if (dotexture)
-                    tex.DefaultTexture.TextureID = textureID;
-                tex.DefaultTexture.RepeatU = (float)scaleU;
-                tex.DefaultTexture.RepeatV = (float)ScaleV;
-                tex.DefaultTexture.OffsetU = (float)offsetU;
-                tex.DefaultTexture.OffsetV = (float)offsetV;
-                tex.DefaultTexture.Rotation = (float)rotation;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        protected void SetTexture(SceneObjectPart part, string texture, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            UUID textureID = ScriptUtils.GetAssetIdFromItemName(m_host, texture, (int)AssetType.Texture);
-            if (textureID.IsZero())
-            {
-                if (!UUID.TryParse(texture, out textureID) || textureID.IsZero())
-                    return;
-            }
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                Primitive.TextureEntryFace texface = tex.CreateFace((uint)face);
-                texface.TextureID = textureID;
-                tex.FaceTextures[face] = texface;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            else if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (uint i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].TextureID = textureID;
-                    }
-                }
-                tex.DefaultTexture.TextureID = textureID;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public void llScaleTexture(double u, double v, int face)
-        {
-
-            ScaleTexture(m_host, u, v, face);
-            ScriptSleep(m_sleepMsOnScaleTexture);
-        }
-
-        protected void ScaleTexture(SceneObjectPart part, double u, double v, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                Primitive.TextureEntryFace texface = tex.CreateFace((uint)face);
-                texface.RepeatU = (float)u;
-                texface.RepeatV = (float)v;
-                tex.FaceTextures[face] = texface;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (int i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].RepeatU = (float)u;
-                        tex.FaceTextures[i].RepeatV = (float)v;
-                    }
-                }
-                tex.DefaultTexture.RepeatU = (float)u;
-                tex.DefaultTexture.RepeatV = (float)v;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public void llOffsetTexture(double u, double v, int face)
-        {
-            OffsetTexture(m_host, u, v, face);
-            ScriptSleep(m_sleepMsOnOffsetTexture);
-        }
-
-        protected void OffsetTexture(SceneObjectPart part, double u, double v, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                Primitive.TextureEntryFace texface = tex.CreateFace((uint)face);
-                texface.OffsetU = (float)u;
-                texface.OffsetV = (float)v;
-                tex.FaceTextures[face] = texface;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (int i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].OffsetU = (float)u;
-                        tex.FaceTextures[i].OffsetV = (float)v;
-                    }
-                }
-                tex.DefaultTexture.OffsetU = (float)u;
-                tex.DefaultTexture.OffsetV = (float)v;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public void llRotateTexture(double rotation, int face)
-        {
-            RotateTexture(m_host, rotation, face);
-            ScriptSleep(m_sleepMsOnRotateTexture);
-        }
-
-        protected void RotateTexture(SceneObjectPart part, double rotation, int face)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face >= 0 && face < nsides)
-            {
-                Primitive.TextureEntryFace texface = tex.CreateFace((uint)face);
-                texface.Rotation = (float)rotation;
-                tex.FaceTextures[face] = texface;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-            if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                for (int i = 0; i < nsides; i++)
-                {
-                    if (tex.FaceTextures[i] != null)
-                    {
-                        tex.FaceTextures[i].Rotation = (float)rotation;
-                    }
-                }
-                tex.DefaultTexture.Rotation = (float)rotation;
-                part.UpdateTextureEntry(tex);
-                return;
-            }
-        }
-
-        public LSL_String llGetTexture(int face)
-        {
-            return GetTexture(m_host, face);
-        }
-
-        protected LSL_String GetTexture(SceneObjectPart part, int face)
-        {
-            Primitive.TextureEntry tex = part.Shape.Textures;
-            int nsides = GetNumberOfSides(part);
-
-            if (face == ScriptBaseClass.ALL_SIDES)
-            {
-                face = 0;
-            }
-
-            if (face >= 0 && face < nsides)
-            {
-                Primitive.TextureEntryFace texface;
-                texface = tex.GetFace((uint)face);
-                string texture = texface.TextureID.ToString();
-
-                lock (part.TaskInventory)
-                {
-                    foreach (KeyValuePair<UUID, TaskInventoryItem> inv in part.TaskInventory)
-                    {
-                        if (inv.Value.AssetID.Equals(texface.TextureID))
-                        {
-                            texture = inv.Value.Name.ToString();
-                            break;
-                        }
-                    }
-                }
-
-                return texture;
-            }
-            else
-            {
-                return ScriptBaseClass.NULL_KEY;
-            }
-        }
-
-        public void llSetPos(LSL_Vector pos)
-        {
-
-            SetPos(m_host, pos, true);
-
-            ScriptSleep(m_sleepMsOnSetPos);
-        }
-
-        /// <summary>
-        /// Tries to move the entire object so that the root prim is within 0.1m of position. http://wiki.secondlife.com/wiki/LlSetRegionPos
-        /// Documentation indicates that the use of x/y coordinates up to 10 meters outside the bounds of a region will work but do not specify what happens if there is no adjacent region for the object to move into.
-        /// Uses the RegionSize constant here rather than hard-coding 266.0 to alert any developer modifying OpenSim to support variable-sized regions that this method will need tweaking.
-        /// </summary>
-        /// <param name="pos"></param>
-        /// <returns>1 if successful, 0 otherwise.</returns>
-        public LSL_Integer llSetRegionPos(LSL_Vector pos)
-        {
-
-            // BEGIN WORKAROUND
-            // IF YOU GET REGION CROSSINGS WORKING WITH THIS FUNCTION, REPLACE THE WORKAROUND.
-            //
-            // This workaround is to prevent silent failure of this function.
-            // According to the specification on the SL Wiki, providing a position outside of the
-            if (pos.x < 0 || pos.x > World.RegionInfo.RegionSizeX || pos.y < 0 || pos.y > World.RegionInfo.RegionSizeY)
-            {
-                return 0;
-            }
-            // END WORK AROUND
-            else if ( // this is not part of the workaround if-block because it's not related to the workaround.
-                IsPhysical() ||
-                m_host.ParentGroup.IsAttachment || // return FALSE if attachment
-                (
-                    pos.x < -10.0 || // return FALSE if more than 10 meters into a west-adjacent region.
-                    pos.x > (World.RegionInfo.RegionSizeX + 10) || // return FALSE if more than 10 meters into a east-adjacent region.
-                    pos.y < -10.0 || // return FALSE if more than 10 meters into a south-adjacent region.
-                    pos.y > (World.RegionInfo.RegionSizeY + 10) || // return FALSE if more than 10 meters into a north-adjacent region.
-                    pos.z > Constants.RegionHeight // return FALSE if altitude than 4096m
-                )
-            )
-            {
-                return 0;
-            }
-
-            // if we reach this point, then the object is not physical, it's not an attachment, and the destination is within the valid range.
-            // this could possibly be done in the above else-if block, but we're doing the check here to keep the code easier to read.
-
-            Vector3 objectPos = m_host.ParentGroup.RootPart.AbsolutePosition;
-            LandData here = World.GetLandData(objectPos);
-            LandData there = World.GetLandData(pos);
-
-            // we're only checking prim limits if it's moving to a different parcel under the assumption that if the object got onto the parcel without exceeding the prim limits.
-
-            bool sameParcel = here.GlobalID.Equals(there.GlobalID);
-
-            if (!sameParcel && !World.Permissions.CanRezObject(
-                m_host.ParentGroup.PrimCount, m_host.ParentGroup.OwnerID, pos))
-            {
-                return 0;
-            }
-
-            SetPos(m_host.ParentGroup.RootPart, pos, false);
-
-            return VecDistSquare(pos, llGetRootPosition()) <= 0.01 ? 1 : 0;
-        }
-
-        // Capped movemment if distance > 10m (http://wiki.secondlife.com/wiki/LlSetPos)
-        // note linked setpos is capped "differently"
-        private LSL_Vector SetPosAdjust(LSL_Vector start, LSL_Vector end)
-        {
-            if (VecDistSquare(start, end) > m_Script10mDistanceSquare)
-                return start + m_Script10mDistance * llVecNorm(end - start);
-            else
-                return end;
-        }
-
-        protected LSL_Vector GetSetPosTarget(SceneObjectPart part, LSL_Vector targetPos, LSL_Vector fromPos, bool adjust)
-        {
-            if (part == null)
-                return targetPos;
-            SceneObjectGroup grp = part.ParentGroup;
-            if (grp == null || grp.IsDeleted || grp.inTransit)
-                return targetPos;
-
-            if (adjust)
-                targetPos = SetPosAdjust(fromPos, targetPos);
-
-            if (m_disable_underground_movement && grp.AttachmentPoint == 0)
-            {
-                if (part.IsRoot)
-                {
-                    float ground = World.GetGroundHeight((float)targetPos.x, (float)targetPos.y);
-                    if ((targetPos.z < ground))
-                        targetPos.z = ground;
-                }
-            }
-            return targetPos;
-        }
-
-        /// <summary>
-        /// set object position, optionally capping the distance.
-        /// </summary>
-        /// <param name="part"></param>
-        /// <param name="targetPos"></param>
-        /// <param name="adjust">if TRUE, will cap the distance to 10m.</param>
-        protected void SetPos(SceneObjectPart part, LSL_Vector targetPos, bool adjust)
-        {
-            if (part == null)
-                return;
-
-            SceneObjectGroup grp = part.ParentGroup;
-            if (grp == null || grp.IsDeleted || grp.inTransit)
-                return;
-
-            LSL_Vector currentPos = GetPartLocalPos(part);
-            LSL_Vector toPos = GetSetPosTarget(part, targetPos, currentPos, adjust);
-
-            if (part.IsRoot)
-            {
-                if (!grp.IsAttachment && !World.Permissions.CanObjectEntry(grp, false, (Vector3)toPos))
-                    return;
-                grp.UpdateGroupPosition((Vector3)toPos);
-            }
-            else
-            {
-                part.OffsetPosition = (Vector3)toPos;
-//                SceneObjectGroup parent = part.ParentGroup;
-//                parent.HasGroupChanged = true;
-//                parent.ScheduleGroupForTerseUpdate();
-                part.ScheduleTerseUpdate();
-            }
-        }
-
-        public LSL_Vector llGetPos()
-        {
-            return m_host.GetWorldPosition();
-        }
-
-        public LSL_Vector llGetLocalPos()
-        {
-            return GetPartLocalPos(m_host);
-        }
-
-        protected LSL_Vector GetPartLocalPos(SceneObjectPart part)
-        {
-            Vector3 pos;
-            if (part.IsRoot)
-            {
-                if (part.ParentGroup.IsAttachment)
-                    pos = part.AttachedPos;
-                else
-                    pos = part.AbsolutePosition;
-            }
-            else
-            {
-                pos = part.OffsetPosition;
-            }
-
-            //m_log.DebugFormat("[LSL API]: Returning {0} in GetPartLocalPos()", pos);
-            return new LSL_Vector(pos);
-        }
-
-        public void llSetRot(LSL_Rotation rot)
-        {
-
-            // try to let this work as in SL...
-            if (m_host.ParentID == 0 || (m_host.ParentGroup != null && m_host == m_host.ParentGroup.RootPart))
-            {
-                // special case: If we are root, rotate complete SOG to new rotation
-                SetRot(m_host, rot);
-            }
-            else
-            {
-                // we are a child. The rotation values will be set to the one of root modified by rot, as in SL. Don't ask.
-                SceneObjectPart rootPart = m_host.ParentGroup.RootPart;
-                if (rootPart != null) // better safe than sorry
-                {
-                    SetRot(m_host, rootPart.RotationOffset * (Quaternion)rot);
-                }
-            }
-
-            ScriptSleep(m_sleepMsOnSetRot);
-        }
-
-        public void llSetLocalRot(LSL_Rotation rot)
-        {
-            SetRot(m_host, rot);
-            ScriptSleep(m_sleepMsOnSetLocalRot);
-        }
-
-        protected void SetRot(SceneObjectPart part, Quaternion rot)
-        {
-            if (part == null || part.ParentGroup == null || part.ParentGroup.IsDeleted)
-                return;
-
-            bool isroot = (part == part.ParentGroup.RootPart);
-            bool isphys;
-
-            PhysicsActor pa = part.PhysActor;
-
-            // keep using physactor ideia of isphysical
-            // it should be SOP ideia of that
-            // not much of a issue with ubOde
-            if (pa != null && pa.IsPhysical)
-                isphys = true;
-            else
-                isphys = false;
-
-            // SL doesn't let scripts rotate root of physical linksets
-            if (isroot && isphys)
-                return;
-
-            part.UpdateRotation(rot);
-
-            // Update rotation does not move the object in the physics engine if it's a non physical linkset
-            // so do a nasty update of parts positions if is a root part rotation
-            if (isroot && pa != null) // with if above implies non physical  root part
-            {
-                part.ParentGroup.ResetChildPrimPhysicsPositions();
-            }
-            else // fix sitting avatars. This is only needed bc of how we link avas to child parts, not root part
-            {
-                //                List<ScenePresence> sittingavas = part.ParentGroup.GetLinkedAvatars();
-                List<ScenePresence> sittingavas = part.ParentGroup.GetSittingAvatars();
-                if (sittingavas.Count > 0)
-                {
-                    foreach (ScenePresence av in sittingavas)
-                    {
-                        if (isroot || part.LocalId == av.ParentID)
-                            av.SendTerseUpdateToAllClients();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// See http://lslwiki.net/lslwiki/wakka.php?wakka=ChildRotation
-        /// </summary>
-        public LSL_Rotation llGetRot()
-        {
-            // unlinked or root prim then use llRootRotation
-            // see llRootRotaion for references.
-            if (m_host.LinkNum == 0 || m_host.LinkNum == 1)
-            {
-                return llGetRootRotation();
-            }
-
-            Quaternion q = m_host.GetWorldRotation();
-
-            if (m_host.ParentGroup != null && m_host.ParentGroup.AttachmentPoint != 0)
-            {
-                ScenePresence avatar = World.GetScenePresence(m_host.ParentGroup.AttachedAvatar);
-                if (avatar != null)
-                {
-                    if ((avatar.AgentControlFlags & (uint)AgentManager.ControlFlags.AGENT_CONTROL_MOUSELOOK) != 0)
-                        q = avatar.CameraRotation * q; // Mouselook
-                    else
-                        q = avatar.Rotation * q; // Currently infrequently updated so may be inaccurate
-                }
-            }
-
-            return new LSL_Rotation(q.X, q.Y, q.Z, q.W);
-        }
-
-        private LSL_Rotation GetPartRot(SceneObjectPart part)
-        {
-            Quaternion q;
-            if (part.LinkNum == 0 || part.LinkNum == 1) // unlinked or root prim
-            {
-                if (part.ParentGroup.AttachmentPoint != 0)
-                {
-                    ScenePresence avatar = World.GetScenePresence(part.ParentGroup.AttachedAvatar);
-                    if (avatar != null)
-                    {
-                        if ((avatar.AgentControlFlags & (uint)AgentManager.ControlFlags.AGENT_CONTROL_MOUSELOOK) != 0)
-                            q = avatar.CameraRotation; // Mouselook
-                        else
-                            q = avatar.GetWorldRotation(); // Currently infrequently updated so may be inaccurate
-                    }
-                    else
-                        q = part.ParentGroup.GroupRotation; // Likely never get here but just in case
-                }
-                else
-                    q = part.ParentGroup.GroupRotation; // just the group rotation
-
-                return new LSL_Rotation(q);
-            }
-
-            q = part.GetWorldRotation();
-            if (part.ParentGroup.AttachmentPoint != 0)
-            {
-                ScenePresence avatar = World.GetScenePresence(part.ParentGroup.AttachedAvatar);
-                if (avatar != null)
-                {
-                    if ((avatar.AgentControlFlags & (uint)AgentManager.ControlFlags.AGENT_CONTROL_MOUSELOOK) != 0)
-                        q = avatar.CameraRotation * q; // Mouselook
-                    else
-                        q = avatar.Rotation * q; // Currently infrequently updated so may be inaccurate
-                }
-            }
-
-            return new LSL_Rotation(q.X, q.Y, q.Z, q.W);
-        }
-
-        public LSL_Rotation llGetLocalRot()
-        {
-            return GetPartLocalRot(m_host);
-        }
-
-        private LSL_Rotation GetPartLocalRot(SceneObjectPart part)
-        {
-            Quaternion rot = part.RotationOffset;
-            return new LSL_Rotation(rot.X, rot.Y, rot.Z, rot.W);
-        }
-
-        public void llSetForce(LSL_Vector force, int local)
-        {
-
-            if (!m_host.ParentGroup.IsDeleted)
-            {
-                if (local != 0)
-                    force *= llGetRot();
-
-                m_host.ParentGroup.RootPart.SetForce(force);
-            }
-        }
-
-        public LSL_Vector llGetForce()
-        {
-            LSL_Vector force = new LSL_Vector(0.0, 0.0, 0.0);
-
-
-            if (!m_host.ParentGroup.IsDeleted)
-            {
-                force = m_host.ParentGroup.RootPart.GetForce();
-            }
-
-            return force;
-        }
-
-        public void llSetVelocity(LSL_Vector vel, int local)
-        {
-            m_host.SetVelocity(new Vector3((float)vel.x, (float)vel.y, (float)vel.z), local != 0);
-        }
-
-        public void llSetAngularVelocity(LSL_Vector avel, int local)
-        {
-            m_host.SetAngularVelocity(new Vector3((float)avel.x, (float)avel.y, (float)avel.z), local != 0);
-        }
-        public LSL_Integer llTarget(LSL_Vector position, double range)
-        {
-            return m_host.ParentGroup.RegisterTargetWaypoint(m_item.ItemID, position, (float)range);
-        }
-
-        public void llTargetRemove(int number)
-        {
-            m_host.ParentGroup.UnregisterTargetWaypoint(number);
-        }
-
-        public LSL_Integer llRotTarget(LSL_Rotation rot, double error)
-        {
-            return m_host.ParentGroup.RegisterRotTargetWaypoint(m_item.ItemID, rot, (float)error);
-        }
-
-        public void llRotTargetRemove(int number)
-        {
-            m_host.ParentGroup.UnRegisterRotTargetWaypoint(number);
-        }
-
-        public void llMoveToTarget(LSL_Vector target, double tau)
-        {
-            m_host.ParentGroup.MoveToTarget(target, (float)tau);
-        }
-
-        public void llStopMoveToTarget()
-        {
-            m_host.ParentGroup.StopMoveToTarget();
-        }
-
-        public void llApplyImpulse(LSL_Vector force, LSL_Integer local)
-        {
-            //No energy force yet
-            Vector3 v = force;
-            if (v.Length() > 20000.0f)
-            {
-                v.Normalize();
-                v = v * 20000.0f;
-            }
-            m_host.ApplyImpulse(v, local != 0);
-        }
-
-
-        public void llApplyRotationalImpulse(LSL_Vector force, int local)
-        {
-            m_host.ParentGroup.RootPart.ApplyAngularImpulse(force, local != 0);
-        }
-
-        public void llSetTorque(LSL_Vector torque, int local)
-        {
-            m_host.ParentGroup.RootPart.SetAngularImpulse(torque, local != 0);
-        }
-
-        public LSL_Vector llGetTorque()
-        {
-
-            return new LSL_Vector(m_host.ParentGroup.GetTorque());
-        }
-
-        public void llSetForceAndTorque(LSL_Vector force, LSL_Vector torque, int local)
-        {
-            llSetForce(force, local);
-            llSetTorque(torque, local);
-        }
-
-
-        public LSL_Vector llGetVel()
-        {
-
-            Vector3 vel = Vector3.Zero;
-
-            if (m_host.ParentGroup.IsAttachment)
-            {
-                ScenePresence avatar = m_host.ParentGroup.Scene.GetScenePresence(m_host.ParentGroup.AttachedAvatar);
-                if (avatar != null)
-                    vel = avatar.GetWorldVelocity();
-            }
-            else
-            {
-                vel = m_host.ParentGroup.RootPart.Velocity;
-            }
-
-            return new LSL_Vector(vel);
-        }
-
-        public LSL_Vector llGetAccel()
-        {
-
-            return new LSL_Vector(m_host.Acceleration);
-        }
-
-        public LSL_Vector llGetOmega()
-        {
-            Vector3 avel = m_host.AngularVelocity;
-            return new LSL_Vector(avel.X, avel.Y, avel.Z);
-        }
-
-        public LSL_Float llGetTimeOfDay()
-        {
-            return (double)((DateTime.Now.TimeOfDay.TotalMilliseconds / 1000) % (3600 * 4));
-        }
-
-        public LSL_Float llGetWallclock()
-        {
-            DateTimeOffset dateTimeOffset = DateTimeOffset.Now;
-
-            if (string.IsNullOrEmpty(m_GetWallclockTimeZone) == false)
-            {
-                dateTimeOffset = 
-                    TimeZoneInfo.ConvertTimeBySystemTimeZoneId(
-                        DateTimeOffset.UtcNow, m_GetWallclockTimeZone);
-            }
-
-            return Math.Truncate(dateTimeOffset.DateTime.TimeOfDay.TotalSeconds);
-        }
-
-        public LSL_Float llGetTime()
-        {
-            double ScriptTime = Util.GetTimeStampMS() - m_timer;
-            return (float)Math.Round((ScriptTime / 1000.0), 3);
-        }
-
-        public void llResetTime()
-        {
-            m_timer = Util.GetTimeStampMS();
-        }
-
-        public LSL_Float llGetAndResetTime()
-        {
-            double now = Util.GetTimeStampMS();
-            double ScriptTime = now - m_timer;
-            m_timer = now;
-            return (float)Math.Round((ScriptTime / 1000.0), 3);
-        }
-
-        public void llSound(string sound, double volume, int queue, int loop)
-        {
-            Deprecated("llSound", "Use llPlaySound instead");
-        }
-
-        // Xantor 20080528 PlaySound updated so it accepts an objectinventory name -or- a key to a sound
-        // 20080530 Updated to remove code duplication
-        public void llPlaySound(string sound, double volume)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            // send the sound, once, to all clients in range
-            m_SoundModule.SendSound(m_host.UUID, soundID, volume, false, 0, false, false);
-        }
-
-        public void llLinkPlaySound(LSL_Integer linknumber, string sound, double volume)
-        {
-            if (m_SoundModule == null)
-                return;
-            if (m_host.ParentGroup == null || m_host.ParentGroup.IsDeleted)
-                return;
-
-            SceneObjectPart sop;
-            if (linknumber == ScriptBaseClass.LINK_THIS)
-                sop = m_host;
-            else if (linknumber < 0)
-                return;
-            else if (linknumber < 2)
-                sop = m_host.ParentGroup.RootPart;
-            else
-                sop = m_host.ParentGroup.GetLinkNumPart(linknumber);
-
-            if(sop == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if (soundID.IsZero())
-                return;
-
-            // send the sound, once, to all clients in range
-            m_SoundModule.SendSound(sop.UUID, soundID, volume, false, 0, false, false);
-        }
-
-        public void llLoopSound(string sound, double volume)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            m_SoundModule.LoopSound(m_host.UUID, soundID, volume, false,false);
-        }
-
-        public void llLoopSoundMaster(string sound, double volume)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            m_SoundModule.LoopSound(m_host.UUID, soundID, volume, true, false);
-        }
-
-        public void llLoopSoundSlave(string sound, double volume)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            m_SoundModule.LoopSound(m_host.UUID, soundID, volume, false, true);
-        }
-
-        public void llPlaySoundSlave(string sound, double volume)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            // send the sound, once, to all clients in range
-            m_SoundModule.SendSound(m_host.UUID, soundID, volume, false, 0, true, false);
-        }
-
-        public void llTriggerSound(string sound, double volume)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            // send the sound, once, to all clients in rangeTrigger or play an attached sound in this part's inventory.
-            m_SoundModule.SendSound(m_host.UUID, soundID, volume, true, 0, false, false);
-        }
-
-        public void llStopSound()
-        {
-
-            if (m_SoundModule != null)
-                m_SoundModule.StopSound(m_host.UUID);
-        }
-
-        public void llLinkStopSound(LSL_Integer linknumber)
-        {
-            if (m_SoundModule != null)
-            {
-                foreach(SceneObjectPart sop in GetLinkParts(linknumber))
-                    m_SoundModule.StopSound(sop.UUID);
-            }
-        }
-
-        public void llPreloadSound(string sound)
-        {
-
-            if (m_SoundModule == null)
-                return;
-
-            UUID soundID = ScriptUtils.GetAssetIdFromKeyOrItemName(m_host, sound, AssetType.Sound);
-            if(soundID.IsZero())
-                return;
-
-            m_SoundModule.PreloadSound(m_host.UUID, soundID);
-            ScriptSleep(m_sleepMsOnPreloadSound);
-        }
 
         /// <summary>
         /// Return a portion of the designated string bounded by
